% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{accumulators}
\alias{accumulators}
\alias{zeronames}
\title{accumulators}
\description{
Accumulator variables
}
\details{
In formulating models, one sometimes wishes to define a state variable that will accumulate some quantity over the interval between successive observations.
\pkg{pomp} provides a facility to make such features more convenient.
Specifically, variables named in the \code{pomp}'s \code{zeronames} argument will be set to zero immediately following each observation.
See \code{\link{sir}} and the tutorials on the \href{https://kingaa.github.io/pomp/}{package website} for examples.
}
\seealso{
Other information on model implementation: \code{\link{Csnippet}},
  \code{\link{covariates}}, \code{\link{dmeasure_plugins}},
  \code{\link{dprocess_plugins}},
  \code{\link{parameter_trans}},
  \code{\link{prior_plugins}}, \code{\link{rinit_plugins}},
  \code{\link{rmeasure_plugins}},
  \code{\link{rprocess_plugins}},
  \code{\link{skeleton_plugins}}
}
\concept{information on model implementation}
