% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pomp.R, R/traj_match.R
\name{as.pomp}
\alias{as.pomp}
\alias{coerce-traj_match_objfun-pomp}
\alias{coerce,traj_match_objfun,pomp-method}
\alias{as.pomp.traj_match_objfun}
\title{as.pomp}
\usage{
as.pomp(object, ...)

\method{as.pomp}{traj_match_objfun}(object)
}
\arguments{
\item{object}{the object to be coerced}

\item{\dots}{additional arguments}
}
\description{
Coerce to a \sQuote{pomp} object
}
\keyword{internal}
