% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake.R
\name{bake}
\alias{bake}
\alias{stew}
\alias{freeze}
\title{Bake, stew, and freeze}
\usage{
bake(
  file,
  expr,
  seed = NULL,
  kind = NULL,
  normal.kind = NULL,
  dependson = NULL
)

stew(
  file,
  expr,
  seed = NULL,
  kind = NULL,
  normal.kind = NULL,
  dependson = NULL
)

freeze(
  expr,
  seed = NULL,
  kind = NULL,
  normal.kind = NULL,
  envir = parent.frame(),
  enclos = if (is.list(envir) || is.pairlist(envir)) parent.frame() else baseenv()
)
}
\arguments{
\item{file}{Name of the binary data file in which the result will be stored or retrieved, as appropriate.
For \code{bake}, this will contain a single object and hence be an RDS file (extension \sQuote{rds});
for \code{stew}, this will contain one or more named objects and hence be an RDA file (extension \sQuote{rda}).}

\item{expr}{Expression to be evaluated.}

\item{seed, kind, normal.kind}{optional.
To set the state and of the RNG.
See \code{\link{set.seed}}.
The default, \code{seed = NULL}, will not change the RNG state.
\code{seed} should be a single integer.
See \code{\link{set.seed}}.}

\item{dependson}{optional.
Variables on which the computation in \code{expr} depends.
Hashes of these objects will be stored in \code{file}, along with the results of evaluation \code{expr}.
When \code{file} exists, hashes of these objects will be compared against the stored values;
recomputation is forced when these do not match.
These objects should be specified as unquoted symbols, using \code{c} or \code{list} if there are more than one.}

\item{envir}{the \code{\link[base]{environment}} in which \code{expr} is to
    be evaluated.  May also be \code{NULL}, a list, a data frame,
    a pairlist or an integer as specified to \code{\link[base]{sys.call}}.}

\item{enclos}{Relevant when \code{envir} is a (pair)list or a data frame.
    Specifies the enclosure, i.e., where \R looks for objects not found
    in \code{envir}.  This can be \code{NULL} (interpreted as the base
    package environment, \code{\link[base]{baseenv}()}) or an environment.}
}
\value{
\code{bake} returns the value of the evaluated expression \code{expr}.
Other objects created in the evaluation of \code{expr} are discarded along with the temporary, local environment created for the evaluation.

The latter behavior differs from that of \code{stew}, which returns the names of the objects created during the evaluation of \code{expr}.
After \code{stew} completes, these objects exist in the parent environment (that from which \code{stew} was called).

\code{freeze} returns the value of evaluated expression \code{expr}.
However, \code{freeze} evaluates \code{expr} within the parent environment, so other objects created in the evaluation of \code{expr} will therefore exist after \code{freeze} completes.

\code{bake} and \code{stew} return information about the time used in evaluating the expression.
This is recorded in the \code{system.time} attribute of the return value.
In addition, if \code{seed} is specified, information about the seed (and the kind of random-number generator used) are stored as attributes of the return value.
}
\description{
Tools for reproducible computations.
}
\details{
On cooking shows, recipes requiring lengthy baking or stewing are prepared beforehand.
The \code{bake} and \code{stew} functions perform analogously:
an computation is performed and stored in a named file.
If the function is called again and the file is present, the computation is not executed.
Instead, the results are loaded from the file in which they were previously stored.
Moreover, via their optional \code{seed} argument, \code{bake} and \code{stew} can control the pseudorandom-number generator (RNG) for greater reproducibility.
After the computation is finished, these functions restore the pre-existing RNG state to avoid side effects.

The \code{freeze} function doesn't save results, but does set the RNG state to the specified value and restore it after the computation is complete.

Both \code{bake} and \code{stew} first test to see whether \code{file} exists.
If it does, \code{bake} reads it using \code{\link{readRDS}} and returns the resulting object.
By contrast, \code{stew} loads the file using \code{\link{load}} and copies the objects it contains into the user's workspace (or the environment of the call to \code{stew}).

If \code{file} does not exist, then both \code{bake} and \code{stew} evaluate the expression \code{expr}; they differ in the results that they save.
\code{bake} saves the value of the evaluated expression to \code{file} as a single object.
The name of that object is not saved.
By contrast, \code{stew} creates a local environment within which \code{expr} is evaluated; all objects in that environment are saved (by name) in \code{file}.
}
\examples{
\dontrun{
  bake(file="example1.rds",{
    x <- runif(1000)
    mean(x)
  })

  bake(file="example1.rds",{
    x <- runif(1000)
    mean(x)
  })

  bake(file="example1.rds",{
    a <- 3
    x <- runif(1000)
    mean(x)
  })

  a <- 5
  b <- 2

  stew(file="example2.rda",
    dependson=c(a,b),{
      x <- runif(10)
      y <- rnorm(n=10,mean=a*x+b,sd=2)
    })

  plot(x,y)

  set.seed(11)
  runif(2)
  freeze(runif(3),seed=5886730)
  runif(2)
  freeze(runif(3),seed=5886730)
  runif(2)

  set.seed(11)
  runif(2)
  runif(2)
  runif(2)

}
}
\author{
Aaron A. King
}
