% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplines.R
\name{bsplines}
\alias{bsplines}
\alias{bspline_basis}
\alias{periodic_bspline_basis}
\title{B-spline bases}
\usage{
bspline_basis(x, nbasis, degree = 3, deriv = 0, names = NULL, rg = range(x))

periodic_bspline_basis(
  x,
  nbasis,
  degree = 3,
  period = 1,
  deriv = 0,
  names = NULL
)
}
\arguments{
\item{x}{Vector at which the spline functions are to be evaluated.}

\item{nbasis}{The number of basis functions to return.}

\item{degree}{Degree of requested B-splines.}

\item{deriv}{The order of the derivative required.}

\item{names}{optional; the names to be given to the basis functions.  These
will be the column-names of the matrix returned.  If the names are
specified as a format string (e.g., "basis\%d"), \code{\link{sprintf}} will
be used to generate the names from the column number.  If a single
non-format string is specified, the names will be generated by
\code{\link{paste}}-ing \code{name} to the column number.  One can also
specify each column name explicitly by giving a length-\code{nbasis} string
vector.  By default, no column-names are given.}

\item{rg}{numeric of length 2; range of the B-spline basis.
To be properly specified, we must have \code{rg[1] < rg[2]}.}

\item{period}{The period of the requested periodic B-splines.}
}
\value{
\item{bspline_basis}{ Returns a matrix with \code{length(x)} rows
and \code{nbasis} columns.  Each column contains the values one of the
spline basis functions.}
\item{periodic_bspline_basis}{ Returns a matrix with \code{length(x)} rows
and \code{nbasis} columns.  The basis functions returned are periodic with
period \code{period}.}
If \code{deriv>0}, the derivative of that order of each of the corresponding spline basis functions are returned.
}
\description{
These functions generate B-spline basis functions.
\code{bspline_basis} gives a basis of spline functions.
\code{periodic_bspline_basis} gives a
basis of periodic spline functions.
}
\section{C API}{

Access to the underlying C routines is available: see
\href{https://kingaa.github.io/pomp/C_API.html}{the \pkg{pomp} C API document}.
for definition and documentation of the C API.
}

\examples{

x <- seq(0,2,by=0.01)
y <- bspline_basis(x,degree=3,nbasis=9,names="basis")
matplot(x,y,type='l',ylim=c(0,1.1))
lines(x,apply(y,1,sum),lwd=2)

x <- seq(-1,2,by=0.01)
y <- periodic_bspline_basis(x,nbasis=5,names="spline\%d")
matplot(x,y,type='l')

}
\seealso{
More on interpolation:
\code{\link{covariates}},
\code{\link[=lookup]{lookup()}}
}
\author{
Aaron A. King
}
\concept{interpolation}
\concept{splines}
\keyword{smooth}
