% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\alias{c}
\alias{c.Pomp}
\title{Combine}
\usage{
\method{c}{Pomp}(...)

combine(...)
}
\arguments{
\item{...}{elements to be recursively combined into a \sQuote{listie}}
}
\value{
\code{combine} applied to one or lists applies \code{c} to convert the list into a \sQuote{listie}.
In particular, \code{combine(A,B,C)} is equivalent to \code{do.call(c,list(A,B,C))}.
}
\description{
Combine two or more \sQuote{pomp} objects into a list-like \sQuote{listie}.
}
\keyword{internal}
