% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc.R
\name{conc}
\alias{conc}
\alias{conc,Pomp-method}
\alias{conc,Pfilter-method}
\alias{conc,Abc-method}
\alias{conc,Mif2-method}
\alias{conc,Pmcmc-method}
\title{Concatenate}
\usage{
\S4method{conc}{Pomp}(...)

\S4method{conc}{Pfilter}(...)

\S4method{conc}{Abc}(...)

\S4method{conc}{Mif2}(...)

\S4method{conc}{Pmcmc}(...)
}
\description{
Internal methods to concatenate objects into useful listie.
}
\details{
Not exported.
}
\keyword{internal}
