% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{conjoin}
\alias{conjoin}
\alias{c}
\alias{c.Pomp}
\title{Conjoin}
\usage{
\method{c}{Pomp}(...)

conjoin(...)
}
\arguments{
\item{...}{elements to be recursively combined into a \sQuote{listie}}
}
\description{
Combine two or more \sQuote{pomp} objects into a list-like \sQuote{listie}.
}
\details{
\code{conjoin} applied to one or lists applies \code{c} to convert the list into a \sQuote{listie}.
In particular, \code{conjoin(A,B,C)} is equivalent to \code{do.call(c,unlist(list(A,B,C)))}.
}
\keyword{internal}
