% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\name{dmeasure-generic}
\alias{dmeasure-generic}
\alias{dmeasure,missing-method}
\alias{dmeasure,ANY-method}
\title{Generic dmeasure}
\usage{
dmeasure(object, ...)
}
\description{
Generic dmeasure
}
\keyword{internal}
