% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\docType{methods}
\name{dmeasure}
\alias{dmeasure}
\alias{dmeasure,ANY-method}
\alias{dmeasure,missing-method}
\alias{dmeasure,pomp-method}
\title{dmeasure workhorse}
\usage{
\S4method{dmeasure}{pomp}(
  object,
  ...,
  y = obs(object),
  x = states(object),
  times = time(object),
  params = coef(object),
  log = FALSE
)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}.
This will typically be the output of \code{pomp}, \code{simulate}, or one of the \pkg{pomp} inference algorithms.}

\item{...}{additional arguments are ignored.}

\item{y}{a matrix containing observations.
The dimensions of \code{y} are \code{nobs} x \code{ntimes}, where \code{nobs} is the number of observables
and \code{ntimes} is the length of \code{times}.}

\item{x}{an array containing states of the unobserved process.
The dimensions of \code{x} are \code{nvars} x \code{nrep} x \code{ntimes},
where \code{nvars} is the number of state variables,
\code{nrep} is the number of replicates,
and \code{ntimes} is the length of \code{times}.
One can also pass \code{x} as a named numeric vector, which is equivalent to the \code{nrep=1}, \code{ntimes=1} case.}

\item{times}{a numeric vector (length \code{ntimes}) containing times.
These must be in non-decreasing order.}

\item{params}{a \code{npar} x \code{nrep} matrix of parameters.
Each column is treated as an independent parameter set, in correspondence with the corresponding column of \code{x}.}

\item{log}{if TRUE, log probabilities are returned.}
}
\value{
\code{dmeasure} returns a matrix of dimensions \code{nreps} x \code{ntimes}.
If \code{d} is the returned matrix, \code{d[j,k]} is the likelihood (or log likelihood if \code{log = TRUE}) of the observation \code{y[,k]} at time \code{times[k]} given the state \code{x[,j,k]}.
}
\description{
\code{dmeasure} evaluates the probability density of observations given states.
}
\seealso{
Specification of the measurement density evaluator: \link{dmeasure_spec}

More on \pkg{pomp} workhorse functions:
\code{\link[=dinit]{dinit()}},
\code{\link[=dprior]{dprior()}},
\code{\link[=dprocess]{dprocess()}},
\code{\link[=emeasure]{emeasure()}},
\code{\link[=flow]{flow()}},
\code{\link[=partrans]{partrans()}},
\code{\link{pomp-package}},
\code{\link[=rinit]{rinit()}},
\code{\link[=rmeasure]{rmeasure()}},
\code{\link[=rprior]{rprior()}},
\code{\link[=rprocess]{rprocess()}},
\code{\link[=skeleton]{skeleton()}},
\code{\link[=vmeasure]{vmeasure()}},
\code{\link{workhorses}}
}
\concept{extending the pomp package}
\concept{low-level interface}
\concept{pomp workhorses}
