% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\name{dprocess-generic}
\alias{dprocess-generic}
\alias{dprocess,missing-method}
\alias{dprocess,ANY-method}
\title{Generic dprocess}
\usage{
dprocess(object, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}.}

\item{...}{additional arguments are ignored.}
}
\description{
Generic dprocess
}
\keyword{internal}
