% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_sample_size.R
\name{eff_sample_size}
\alias{eff_sample_size}
\alias{eff_sample_size,missing-method}
\alias{eff_sample_size,ANY-method}
\alias{eff_sample_size,bsmcd_pomp-method}
\alias{eff_sample_size,pfilterd_pomp-method}
\alias{eff_sample_size,wpfilterd_pomp-method}
\alias{eff_sample_size,pfilterList-method}
\title{Effective sample size}
\usage{
\S4method{eff_sample_size}{bsmcd_pomp}(object, ..., format = c("numeric", "data.frame"))

\S4method{eff_sample_size}{pfilterd_pomp}(object, ..., format = c("numeric", "data.frame"))

\S4method{eff_sample_size}{wpfilterd_pomp}(object, ..., format = c("numeric", "data.frame"))

\S4method{eff_sample_size}{pfilterList}(object, ..., format = c("numeric", "data.frame"))
}
\arguments{
\item{object}{result of a filtering computation}

\item{...}{ignored}

\item{format}{format of the returned object}
}
\description{
Estimate the effective sample size of a Monte Carlo computation.
}
\details{
Effective sample size is computed as
\deqn{\left(\sum_i\!w_{it}^2\right)^{-1},}{1/(sum(w_it^2)),}
where \eqn{w_{it}}{w_it} is the normalized weight of particle \eqn{i} at time \eqn{t}.
}
\seealso{
More on sequential Monte Carlo methods:
\code{\link[=bsmc2]{bsmc2()}},
\code{\link[=cond_logLik]{cond_logLik()}},
\code{\link[=filter_mean]{filter_mean()}},
\code{\link[=filter_traj]{filter_traj()}},
\code{\link{kalman}},
\code{\link[=mif2]{mif2()}},
\code{\link[=pfilter]{pfilter()}},
\code{\link[=pmcmc]{pmcmc()}},
\code{\link[=pred_mean]{pred_mean()}},
\code{\link[=pred_var]{pred_var()}},
\code{\link[=saved_states]{saved_states()}},
\code{\link[=wpfilter]{wpfilter()}}

Other extraction methods:
\code{\link[=coef]{coef()}},
\code{\link[=cond_logLik]{cond_logLik()}},
\code{\link[=covmat]{covmat()}},
\code{\link[=filter_mean]{filter_mean()}},
\code{\link[=filter_traj]{filter_traj()}},
\code{\link[=forecast]{forecast()}},
\code{\link[=logLik]{logLik()}},
\code{\link[=obs]{obs()}},
\code{\link[=pred_mean]{pred_mean()}},
\code{\link[=pred_var]{pred_var()}},
\code{\link[=saved_states]{saved_states()}},
\code{\link[=spy]{spy()}},
\code{\link[=states]{states()}},
\code{\link[=summary]{summary()}},
\code{\link[=time]{time()}},
\code{\link[=timezero]{timezero()}},
\code{\link[=traces]{traces()}}
}
\concept{extraction methods}
\concept{particle filter methods}
