% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listie.R, R/extract.R
\docType{class}
\name{listie}
\alias{listie}
\alias{pompList-class}
\alias{abcList-class}
\alias{mif2List-class}
\alias{pmcmcList-class}
\alias{pfilterList-class}
\alias{Pomp-class}
\alias{Pfilter-class}
\alias{Abc-class}
\alias{Mif2-class}
\alias{Pmcmc-class}
\alias{listie-class}
\alias{[-listie}
\alias{[,listie-method}
\title{listie}
\usage{
\S4method{[}{listie}(x, i, j, ..., drop = TRUE)
}
\description{
List-like objects.
}
\keyword{internal}
