% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listies.R, R/extract.R
\docType{class}
\name{listies}
\alias{listies}
\alias{pompList-class}
\alias{abcList-class}
\alias{mif2List-class}
\alias{pmcmcList-class}
\alias{pfilterList-class}
\alias{Pomp-class}
\alias{Pfilter-class}
\alias{Abc-class}
\alias{Mif2-class}
\alias{Pmcmc-class}
\alias{listies-class}
\alias{[-listies}
\alias{[,listies-method}
\title{listies}
\description{
List-like objects.
}
\keyword{internal}
