% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{melt}
\alias{melt}
\alias{melt,ANY-method}
\alias{melt,missing-method}
\alias{melt,array-method}
\alias{melt,list-method}
\title{Melt}
\usage{
\S4method{melt}{ANY}(data, ...)

\S4method{melt}{array}(data, ...)

\S4method{melt}{list}(data, ..., level = 1)
}
\arguments{
\item{data}{object to convert}

\item{...}{ignored}
}
\description{
Convert arrays, lists, and other objects to data frames.
}
\details{
\code{melt} converts its first argument to a data frame.
It is a simplified version of the \code{melt} command provided by the no-longer maintained \pkg{reshape2} package.

An array can be melted into a data frame.
In this case, the data frame will have one row per entry in the array.

A list can be melted into a data frame.
This operation is recursive.
A variable will be appended to distinguish the separate list entries.
}
\keyword{internal}
