% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\name{partrans-generic}
\alias{partrans-generic}
\alias{partrans,missing-method}
\alias{partrans,ANY-method}
\title{Generic partrans}
\usage{
partrans(object, ...)
}
\description{
Generic partrans
}
\keyword{internal}
