% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\docType{import}
\name{pipe}
\alias{pipe}
\alias{\%>\%}
\title{Objects exported from other packages}
\details{
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

