% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/bsmc2.R, R/probe_match.R,
%   R/spect_match.R
\name{plot}
\alias{plot}
\alias{plot,missing-method}
\alias{plot-pomp}
\alias{plot,pomp_plottable-method}
\alias{plot,pomp-method}
\alias{plot,pfilterd_pomp-method}
\alias{plot-pfilterd_pomp}
\alias{plot-Pmcmc}
\alias{plot,Pmcmc-method}
\alias{plot,pmcmcd_pomp-method}
\alias{plot,pmcmcList-method}
\alias{plot-Abc}
\alias{plot,Abc-method}
\alias{plot,abcd_pomp-method}
\alias{plot,abcList-method}
\alias{plot-Mif2}
\alias{plot,Mif2-method}
\alias{plot,mif2d_pomp-method}
\alias{plot,mif2List-method}
\alias{plot-probed_pomp}
\alias{plot,probed_pomp-method}
\alias{plot-spectd_pomp}
\alias{plot,spectd_pomp-method}
\alias{plot-bsmcd_pomp}
\alias{plot,bsmcd_pomp-method}
\alias{plot-probe_match_objfun}
\alias{plot,probe_match_objfun-method}
\alias{plot-spect_match_objfun}
\alias{plot,spect_match_objfun-method}
\title{Plotting}
\usage{
\S4method{plot}{pomp_plottable}(
  x,
  variables,
  panel = lines,
  nc = NULL,
  yax.flip = FALSE,
  mar = c(0, 5.1, 0, if (yax.flip) 5.1 else 2.1),
  oma = c(6, 0, 5, 0),
  axes = TRUE,
  ...
)

\S4method{plot}{Pmcmc}(x, ..., pars)

\S4method{plot}{Abc}(x, ..., pars, scatter = FALSE)

\S4method{plot}{Mif2}(x, y, ...)

\S4method{plot}{probed_pomp}(x, y, ...)

\S4method{plot}{spectd_pomp}(
  x,
  ...,
  max.plots.per.page = 4,
  plot.data = TRUE,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  quantile.styles = list(lwd = 1, lty = 1, col = "gray70"),
  data.styles = list(lwd = 2, lty = 2, col = "black")
)

\S4method{plot}{bsmcd_pomp}(x, pars, thin, ...)

\S4method{plot}{probe_match_objfun}(x, y, ...)

\S4method{plot}{spect_match_objfun}(x, y, ...)
}
\arguments{
\item{x}{the object to plot}

\item{variables}{optional character; names of variables to be displayed}

\item{panel}{function of prototype \code{panel(x, col, bg, pch, type, ...)} which gives the action to be carried out in each panel of the display.}

\item{nc}{the number of columns to use.
Defaults to 1 for up to 4 series, otherwise to 2.}

\item{yax.flip}{logical;
if TRUE, the y-axis (ticks and numbering) should flip from side 2 (left) to 4 (right) from series to series.}

\item{mar, oma}{the \code{\link{par}} \code{mar} and \code{oma} settings.
Modify with care!}

\item{axes}{logical; indicates if x- and y- axes should be drawn}

\item{\dots}{ignored or passed to low-level plotting functions}

\item{pars}{names of parameters.}

\item{scatter}{logical; if \code{FALSE}, traces of the parameters named in \code{pars} will be plotted against ABC iteration number.
If \code{TRUE}, the traces will be displayed or as a scatterplot.}

\item{y}{ignored}

\item{max.plots.per.page}{positive integer; maximum number of plots on a page}

\item{plot.data}{logical; should the data spectrum be included?}

\item{quantiles}{numeric; quantiles to display}

\item{quantile.styles}{list; plot styles to use for quantiles}

\item{data.styles}{list; plot styles to use for data}

\item{thin}{integer; when the number of samples is very large, it can be helpful to plot a random subsample:
\code{thin} specifies the size of this subsample.}
}
\description{
Diagnostic plots.
}
