% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{as.pomp}
\alias{bspline.basis}
\alias{cond.logLik}
\alias{eff.sample.size}
\alias{filter.mean}
\alias{filter.traj}
\alias{mvn.diag.rw}
\alias{mvn.rw}
\alias{mvn.rw.adaptive}
\alias{periodic.bspline.basis}
\alias{pred.mean}
\alias{pred.var}
\alias{probe.acf}
\alias{probe.ccf}
\alias{probe.marginal}
\alias{probe.mean}
\alias{probe.median}
\alias{probe.nlar}
\alias{probe.period}
\alias{probe.quantile}
\alias{probe.sd}
\alias{probe.var}
\alias{rw.sd}
\alias{saved.states}
\title{Defunct functions}
\usage{
as.pomp(...)

bspline.basis(...)

cond.logLik(...)

eff.sample.size(...)

filter.mean(...)

filter.traj(...)

mvn.diag.rw(...)

mvn.rw(...)

mvn.rw.adaptive(...)

periodic.bspline.basis(...)

pred.mean(...)

pred.var(...)

probe.acf(...)

probe.ccf(...)

probe.marginal(...)

probe.mean(...)

probe.median(...)

probe.nlar(...)

probe.period(...)

probe.quantile(...)

probe.sd(...)

probe.var(...)

rw.sd(...)

saved.states(...)
}
\arguments{
\item{...}{all arguments are ignored.}
}
\description{
These functions are defunct and will be expunged in a future release.
}
\keyword{internal}
