% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomp_fun.R
\docType{methods}
\name{pomp_fun}
\alias{pomp_fun}
\alias{pompfunmode}
\title{The "pomp_fun" class}
\format{An object of class \code{list} of length 4.}
\usage{
pompfunmode
}
\arguments{
\item{f}{A function or the name of a native routine.}

\item{PACKAGE}{optional; the name of the dynamically-loadable library in
which the native function \code{f} can be found.}

\item{proto}{optional string; a prototype against which \code{f} will be
checked.}

\item{object, x}{the \sQuote{pomp_fun} object.}
}
\description{
Definition and methods of the \sQuote{pomp_fun} class
}
\details{
The \sQuote{pomp_fun} class helps to settle common issues associated with
user-defined functions which can be defined either via R code or by a
native, compiled routine.  It is not exported to userland.
}
\seealso{
\code{\link{pomp}}
}
\author{
Aaron A. King
}
\keyword{internal}
\keyword{programming}
