% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomp_stop.R
\name{pomp_stop}
\alias{pomp_stop}
\title{pomp_stop}
\usage{
pomp_stop(fn = NULL, ...)
}
\arguments{
\item{fn}{name of function}

\item{\dots}{message}
}
\description{
Custom error function
}
\keyword{internal}
