% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probevals.R
\docType{methods}
\name{probevals}
\alias{probevals}
\alias{probevals,missing-method}
\alias{probevals,ANY-method}
\alias{probevals-probed_pomp}
\alias{probevals,probed_pomp-method}
\title{Probe values}
\usage{
\S4method{probevals}{probed_pomp}(object, ...)
}
\arguments{
\item{object}{an object of class \sQuote{probed_pomp}, or of a class that extends this.}

\item{\dots}{ignored}
}
\description{
Return the values of the probes, evaluated on simulations, and on the data.
}
\seealso{
Other summary statistics methods: \code{\link{abc}},
  \code{\link{basic_probes}}, \code{\link{probe.match}},
  \code{\link{probe}}, \code{\link{spect}}
}
\concept{summary statistics methods}
