% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstop.R
\name{pStop}
\alias{pStop}
\title{pStop}
\usage{
pStop(fn, ...)
}
\arguments{
\item{fn}{name of function (will be enclosed in single quotes)}

\item{\dots}{message}
}
\description{
Custom error function
}
\keyword{internal}
