% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{systematic_resample}
\title{Resample}
\usage{
systematic_resample(weights)
}
\value{
A vector of integers containing the indices of the resample.
}
\description{
Systematic resampling.
}
\keyword{internal}
