% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\name{rmeasure-generic}
\alias{rmeasure-generic}
\alias{rmeasure,missing-method}
\alias{rmeasure,ANY-method}
\title{Generic rmeasure}
\usage{
rmeasure(object, ...)
}
\description{
Generic rmeasure
}
\keyword{internal}
