% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\name{rprocess-generic}
\alias{rprocess-generic}
\alias{rprocess,missing-method}
\alias{rprocess,ANY-method}
\title{Generic rprocess}
\usage{
rprocess(object, ...)
}
\description{
Generic rprocess
}
\keyword{internal}
