% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mif2.R
\name{rw_sd}
\alias{rw_sd}
\title{rw_sd}
\usage{
rw_sd(...)
}
\arguments{
\item{\dots}{Specification of the random-walk intensities (as standard deviations).}
}
\description{
Specifying random-walk intensities.
}
\details{
See \code{\link{mif2}} for details.
}
\seealso{
\code{\link{mif2}}
}
