% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmc2_methods.R
\docType{methods}
\name{Prior and posterior samples}
\alias{Prior and posterior samples}
\alias{prior_samples}
\alias{prior_samples,missing-method}
\alias{prior_samples,ANY-method}
\alias{posterior_samples}
\alias{posterior_samples,missing-method}
\alias{posterior_samples,ANY-method}
\alias{prior_samples-bsmcd_pomp}
\alias{prior_samples,bsmcd_pomp-method}
\alias{posterior_samples-bsmcd_pomp}
\alias{posterior_samples,bsmcd_pomp-method}
\title{Sampling the prior and posterior distributions}
\usage{
\S4method{prior_samples}{bsmcd_pomp}(object, pars, ...)

\S4method{posterior_samples}{bsmcd_pomp}(object, pars, ...)
}
\arguments{
\item{object}{the result of a \code{bsmc2} computation}

\item{pars}{names of parameters}

\item{\dots}{ignored}
}
\description{
These functions return the samples of the prior and posterior distributions
from a \code{bsmc2} run.
}
