% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\name{skeleton-generic}
\alias{skeleton-generic}
\alias{skeleton,missing-method}
\alias{skeleton,ANY-method}
\title{Generic skeleton}
\usage{
skeleton(object, ...)
}
\description{
Generic skeleton
}
\keyword{internal}
