% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy.R
\name{spy}
\alias{spy}
\alias{spy,missing-method}
\alias{spy,ANY-method}
\alias{spy,pomp-method}
\title{Spy}
\usage{
\S4method{spy}{pomp}(object)
}
\arguments{
\item{object}{the object whose structure we wish to examine}
}
\description{
Peek into the inside of one of \pkg{pomp}'s objects.
}
\examples{
\donttest{

  ricker() |> spy()

  sir() |> spy()

  sir2() |> spy()

}
}
\seealso{
Csnippet

Other extraction methods:
\code{\link[=coef]{coef()}},
\code{\link[=cond_logLik]{cond_logLik()}},
\code{\link[=covmat]{covmat()}},
\code{\link[=eff_sample_size]{eff_sample_size()}},
\code{\link[=filter_mean]{filter_mean()}},
\code{\link[=filter_traj]{filter_traj()}},
\code{\link[=forecast]{forecast()}},
\code{\link[=logLik]{logLik()}},
\code{\link[=obs]{obs()}},
\code{\link[=pred_mean]{pred_mean()}},
\code{\link[=pred_var]{pred_var()}},
\code{\link[=saved_states]{saved_states()}},
\code{\link[=states]{states()}},
\code{\link[=summary]{summary()}},
\code{\link[=time]{time()}},
\code{\link[=timezero]{timezero()}},
\code{\link[=traces]{traces()}}
}
\concept{extraction methods}
