% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{methods}
\name{states}
\alias{states}
\alias{states,ANY-method}
\alias{states,missing-method}
\alias{states,pomp-method}
\alias{states,listie-method}
\title{Latent states}
\usage{
\S4method{states}{pomp}(object, vars, ..., format = c("array", "data.frame"))

\S4method{states}{listie}(object, vars, ..., format = c("array", "data.frame"))
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{names of variables to retrieve}

\item{...}{ignored}

\item{format}{format of the returned object}
}
\description{
Extract the latent states from a \sQuote{pomp} object.
}
\seealso{
Other extraction methods:
\code{\link[=coef]{coef()}},
\code{\link[=cond_logLik]{cond_logLik()}},
\code{\link[=covmat]{covmat()}},
\code{\link[=eff_sample_size]{eff_sample_size()}},
\code{\link[=filter_mean]{filter_mean()}},
\code{\link[=filter_traj]{filter_traj()}},
\code{\link[=forecast]{forecast()}},
\code{\link[=logLik]{logLik()}},
\code{\link[=obs]{obs()}},
\code{\link[=pred_mean]{pred_mean()}},
\code{\link[=pred_var]{pred_var()}},
\code{\link[=saved_states]{saved_states()}},
\code{\link[=spy]{spy()}},
\code{\link[=summary]{summary()}},
\code{\link[=time]{time()}},
\code{\link[=timezero]{timezero()}},
\code{\link[=traces]{traces()}}
}
\concept{extraction methods}
