% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R, R/probe.R, R/spect.R, R/objfun.R
\name{summary}
\alias{summary}
\alias{summary,probed_pomp-method}
\alias{summary,spectd_pomp-method}
\alias{summary,objfun-method}
\title{Summary methods}
\usage{
\S4method{summary}{probed_pomp}(object, ...)

\S4method{summary}{spectd_pomp}(object, ...)

\S4method{summary}{objfun}(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{\dots}{ignored or passed to the more primitive function}
}
\description{
Display a summary of a fitted model object.
}
\seealso{
Other extraction methods:
\code{\link[=coef]{coef()}},
\code{\link[=cond_logLik]{cond_logLik()}},
\code{\link[=covmat]{covmat()}},
\code{\link[=eff_sample_size]{eff_sample_size()}},
\code{\link[=filter_mean]{filter_mean()}},
\code{\link[=filter_traj]{filter_traj()}},
\code{\link[=forecast]{forecast()}},
\code{\link[=logLik]{logLik()}},
\code{\link[=obs]{obs()}},
\code{\link[=pred_mean]{pred_mean()}},
\code{\link[=pred_var]{pred_var()}},
\code{\link[=saved_states]{saved_states()}},
\code{\link[=spy]{spy()}},
\code{\link[=states]{states()}},
\code{\link[=time]{time()}},
\code{\link[=timezero]{timezero()}},
\code{\link[=traces]{traces()}}
}
\concept{extraction methods}
