% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\alias{time<-}
\alias{time,missing-method}
\alias{time,pomp-method}
\alias{time<-,pomp-method}
\alias{time,listie-method}
\title{Methods to extract and manipulate the obseration times}
\usage{
\S4method{time}{pomp}(x, t0 = FALSE, ...)

\S4method{time}{pomp}(object, t0 = FALSE, ...) <- value

\S4method{time}{listie}(x, t0 = FALSE, ...)
}
\arguments{
\item{x}{a \sQuote{pomp} object}

\item{t0}{logical; should the zero time be included?}

\item{...}{ignored or passed to the more primitive function}

\item{object}{a \sQuote{pomp} object}

\item{value}{numeric vector; the new vector of times}
}
\description{
Get and set the vector of observation times.
}
\details{
\code{time(object)} returns the vector of observation times.
\code{time(object,t0=TRUE)} returns the vector of observation
times with the zero-time \code{t0} prepended.

\code{time(object) <- value} replaces the observation times slot (\code{times}) of \code{object} with \code{value}.
\code{time(object,t0=TRUE) <- value} has the same effect, but the first element in \code{value} is taken to be the initial time.
The second and subsequent elements of \code{value} are taken to be the observation times.
Those data and states (if they exist) corresponding to the new times are retained.
}
\seealso{
Other extraction methods:
\code{\link[=coef]{coef()}},
\code{\link[=cond_logLik]{cond_logLik()}},
\code{\link[=covmat]{covmat()}},
\code{\link[=eff_sample_size]{eff_sample_size()}},
\code{\link[=filter_mean]{filter_mean()}},
\code{\link[=filter_traj]{filter_traj()}},
\code{\link[=forecast]{forecast()}},
\code{\link[=logLik]{logLik()}},
\code{\link[=obs]{obs()}},
\code{\link[=pred_mean]{pred_mean()}},
\code{\link[=pred_var]{pred_var()}},
\code{\link[=saved_states]{saved_states()}},
\code{\link[=spy]{spy()}},
\code{\link[=states]{states()}},
\code{\link[=summary]{summary()}},
\code{\link[=timezero]{timezero()}},
\code{\link[=traces]{traces()}}
}
\concept{extraction methods}
