% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezero.R
\docType{methods}
\name{timezero}
\alias{timezero}
\alias{timezero<-}
\alias{timezero,missing-method}
\alias{timezero,ANY-method}
\alias{timezero<-,missing-method}
\alias{timezero<-,ANY-method}
\alias{timezero,pomp-method}
\alias{timezero<-,pomp-method}
\title{The zero time}
\usage{
\S4method{timezero}{pomp}(object, ...)

\S4method{timezero}{pomp}(object, ...) <- value
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}}

\item{\dots}{ignored or passed to the more primitive function}

\item{value}{numeric; the new zero-time value}
}
\value{
the value of the zero time
}
\description{
Get and set the zero-time.
}
\seealso{
Other extraction methods:
\code{\link[=coef]{coef()}},
\code{\link[=cond_logLik]{cond_logLik()}},
\code{\link[=covmat]{covmat()}},
\code{\link[=eff_sample_size]{eff_sample_size()}},
\code{\link[=filter_mean]{filter_mean()}},
\code{\link[=filter_traj]{filter_traj()}},
\code{\link[=forecast]{forecast()}},
\code{\link[=logLik]{logLik()}},
\code{\link[=obs]{obs()}},
\code{\link[=pred_mean]{pred_mean()}},
\code{\link[=pred_var]{pred_var()}},
\code{\link[=saved_states]{saved_states()}},
\code{\link[=spy]{spy()}},
\code{\link[=states]{states()}},
\code{\link[=summary]{summary()}},
\code{\link[=time]{time()}},
\code{\link[=traces]{traces()}}
}
\concept{extraction methods}
