% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undefined.R, R/pomp_fun.R, R/skeleton_spec.R,
%   R/rprocess_spec.R
\docType{methods}
\name{undefined}
\alias{undefined}
\alias{undefined,pomp_fun-method}
\alias{undefined,skelPlugin-method}
\alias{undefined,rprocPlugin-method}
\title{Undefined}
\usage{
\S4method{undefined}{pomp_fun}(object, ...)

\S4method{undefined}{skelPlugin}(object, ...)

\S4method{undefined}{rprocPlugin}(object, ...)
}
\arguments{
\item{object}{object to test.}

\item{...}{currently ignored.}
}
\value{
Returns \code{TRUE} if the \pkg{pomp} workhorse method is undefined,
\code{FALSE} if it is defined,
and \code{NA} if the question is inapplicable.
}
\description{
Check for undefined methods.
}
