% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\docType{methods}
\name{window}
\alias{window}
\alias{window-pomp}
\alias{window,pomp-method}
\title{Window}
\usage{
\S4method{window}{pomp}(x, start, end, ...)
}
\arguments{
\item{x}{a \sQuote{pomp} object or object of class extending \sQuote{pomp}}

\item{start, end}{the left and right ends of the window, in units of time}

\item{\dots}{ignored}
}
\description{
Restrict to a portion of a time series.
}
