% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\docType{methods}
\name{abc}
\alias{abc}
\alias{abc,ANY-method}
\alias{abc,missing-method}
\alias{abc-pomp}
\alias{abc,pomp-method}
\alias{abc,probed_pomp-method}
\alias{abc,abcd_pomp-method}
\title{Approximate Bayesian computation}
\usage{
\S4method{abc}{pomp}(object, Nabc = 1, start, proposal, probes, scale,
  epsilon, ..., verbose = getOption("verbose", FALSE))

\S4method{abc}{probed_pomp}(object, probes, ...,
  verbose = getOption("verbose", FALSE))

\S4method{abc}{abcd_pomp}(object, Nabc, start, proposal, probes, scale,
  epsilon, verbose = getOption("verbose"), ...)
}
\arguments{
\item{object}{an object of class \dQuote{pomp}.}

\item{Nabc}{the number of ABC iterations to perform.}

\item{start}{named numeric vector; the starting guess of the parameters.}

\item{proposal}{optional function that draws from the proposal distribution.
Currently, the proposal distribution must be symmetric for proper inference:
it is the user's responsibility to ensure that it is.
Several functions that construct appropriate proposal function are provided:
see \link{MCMC proposals} for more information.}

\item{probes}{List of probes (AKA summary statistics).
See \code{\link{probe}} for details.}

\item{scale}{named numeric vector of scales.}

\item{epsilon}{ABC tolerance.}

\item{\dots}{Additional arguments are passed to \code{\link{pomp}},
allowing one to supply new or modify existing model characteristics or components.}

\item{verbose}{logical; if TRUE, print progress reports.}
}
\description{
The approximate Bayesian computation (ABC) algorithm for estimating the parameters of a partially-observed Markov process.
}
\section{Running ABC}{


\code{abc} returns an object of class \sQuote{abcd_pomp}.
One or more \sQuote{abcd_pomp} objects can be joined to form an \sQuote{abcList} object.
}

\section{Re-running ABC iterations}{


To re-run a sequence of ABC iterations, one can use the \code{abc} method on a \sQuote{abcd_pomp} object.
By default, the same parameters used for the original ABC run are re-used (except for \code{tol}, \code{max.fail}, and \code{verbose}, the defaults of which are shown above).
If one does specify additional arguments, these will override the defaults.
}

\section{Continuing ABC iterations}{


One can continue a series of ABC iterations from where one left off using the \code{continue} method.
A call to \code{abc} to perform \code{Nabc=m} iterations followed by a call to \code{continue} to perform \code{Nabc=n} iterations will produce precisely the same effect as a single call to \code{abc} to perform \code{Nabc=m+n} iterations.
By default, all the algorithmic parameters are the same as used in the original call to \code{abc}.
Additional arguments will override the defaults.
}

\section{ABC diagnostics}{

\code{plot} applied to an ABC object will produce a series of
diagnostic plots.
}

\references{
J.-M. Marin, P. Pudlo, C. P. Robert, and R. J. Ryder,
Approximate Bayesian computational methods.  Statistics and Compuing
22:1167--1180, 2012.

T. Toni and M. P. H. Stumpf, Simulation-based model selection for dynamical
systems in systems and population biology, Bioinformatics 26:104--110,
2010.

T. Toni, D. Welch, N. Strelkowa, A. Ipsen, and M. P. H. Stumpf, Approximate
Bayesian computation scheme for parameter inference and model selection in
dynamical systems Journal of the Royal Society, Interface 6:187--202, 2009.
}
\seealso{
\code{\link{probe}}, \link{MCMC proposals}, and
the tutorials on the \href{https://kingaa.github.io/pomp/}{package website}.

Other summary statistics: \code{\link{Basic probes}},
  \code{\link{Probe matching}}, \code{\link{Probe values}},
  \code{\link{Probes}}
}
\author{
Edward L. Ionides, Aaron A. King
}
\concept{summary statistics}
