% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\docType{methods}
\name{abc}
\alias{abc}
\alias{abc,ANY-method}
\alias{abc,missing-method}
\alias{abc-data.frame}
\alias{abc,data.frame-method}
\alias{abc-pomp}
\alias{abc,pomp-method}
\alias{abc,probed_pomp-method}
\alias{abc,abcd_pomp-method}
\title{Approximate Bayesian computation}
\usage{
\S4method{abc}{data.frame}(data, Nabc = 1, params, rinit, rprocess,
  rmeasure, dprior, proposal, probes, scale, epsilon, ...,
  verbose = getOption("verbose", FALSE))

\S4method{abc}{pomp}(data, Nabc = 1, proposal, probes, scale, epsilon,
  ..., verbose = getOption("verbose", FALSE))

\S4method{abc}{probed_pomp}(data, probes, ...,
  verbose = getOption("verbose", FALSE))

\S4method{abc}{abcd_pomp}(data, Nabc, proposal, probes, scale, epsilon,
  ..., verbose = getOption("verbose", FALSE))
}
\arguments{
\item{data}{either a data frame holding the time series data, or an object of class \sQuote{pomp}, i.e., the output of one of \pkg{pomp}'s methods.}

\item{Nabc}{the number of ABC iterations to perform.}

\item{params}{optional; named numeric vector of parameters.
This will be coerced internally to storage mode \code{double}.}

\item{rinit}{simulator of the initial-state distribution.
This can be furnished either as a C snippet, an \R function, or the name of a pre-compiled native routine available in a dynamically loaded library.
Setting \code{rinit=NULL} sets the initial-state simulator to its default.
For more information, see \link[=rinit_spec]{here}.}

\item{rprocess}{simulator of the latent state process, specified using one of the \link[=rprocess_spec]{rprocess plugins}.
Setting \code{rprocess=NULL} removes the latent-state simulator.
For more information, \link[=rprocess_spec]{see the documentation on these plugins}.}

\item{rmeasure}{simulator of the measurement model, specified either as a C snippet, an \R function, or the name of a pre-compiled native routine available in a dynamically loaded library.
Setting \code{rmeasure=NULL} removes the measurement model simulator.
For more information, see \link[=rmeasure_spec]{here}.}

\item{dprior}{optional; prior distribution density evaluator, specified either as a C snippet, an \R function, or the name of a pre-compiled native routine available in a dynamically loaded library.
Setting \code{dprior=NULL} removes the prior distribution density evaluator.
For more information, see \link[=prior_spec]{here}.}

\item{proposal}{optional function that draws from the proposal
distribution.  Currently, the proposal distribution must be symmetric for
proper inference: it is the user's responsibility to ensure that it is.
Several functions that construct appropriate proposal function are
provided: see \link[=proposals]{MCMC proposals} for more information.}

\item{probes}{a single probe or a list of one or more probes.
A probe is simply a scalar- or vector-valued function of one argument that can be applied to the data array of a \sQuote{pomp}.
A vector-valued probe must always return a vector of the same size.
A number of useful probes are provided with the package:
see \link[=basic_probes]{basic probes}.}

\item{scale}{named numeric vector of scales.}

\item{epsilon}{ABC tolerance.}

\item{...}{additional arguments supply new or modify existing model characteristics or components for use by \pkg{pomp} package algorithms.
See \code{\link{pomp}} for a full list of recognized arguments.
When named arguments not recognized by \code{\link{pomp}} are provided, these are made available to all basic components via the so-called \dfn{userdata} facility.
This allows the user to pass information to the basic components outside of the usual route of covariates (\code{covar}) and model parameters (\code{params}).
See the \link[=userdata]{userdata documentation here} for information on how to use this facility.}

\item{verbose}{logical; if \code{TRUE}, diagnostic messages will be printed to the console.}
}
\description{
The approximate Bayesian computation (ABC) algorithm for estimating the parameters of a partially-observed Markov process.
}
\section{Running ABC}{


\code{abc} returns an object of class \sQuote{abcd_pomp}.
One or more \sQuote{abcd_pomp} objects can be joined to form an \sQuote{abcList} object.
}

\section{Re-running ABC iterations}{


To re-run a sequence of ABC iterations, one can use the \code{abc} method on a \sQuote{abcd_pomp} object.
By default, the same parameters used for the original ABC run are re-used (except for \code{tol}, \code{max.fail}, and \code{verbose}, the defaults of which are shown above).
If one does specify additional arguments, these will override the defaults.
}

\section{Continuing ABC iterations}{


One can continue a series of ABC iterations from where one left off using the \code{continue} method.
A call to \code{abc} to perform \code{Nabc=m} iterations followed by a call to \code{continue} to perform \code{Nabc=n} iterations will produce precisely the same effect as a single call to \code{abc} to perform \code{Nabc=m+n} iterations.
By default, all the algorithmic parameters are the same as used in the original call to \code{abc}.
Additional arguments will override the defaults.
}

\section{Methods}{

The following can be applied to the output of an \code{abc} operation:
\describe{
\item{plot}{produces a series of diagnostic plots}
\item{traces}{produces a \code{\link[coda]{mcmc}} object, to which the various \pkg{coda} convergence diagnostics can be applied}
}
}

\references{
J.-M. Marin, P. Pudlo, C. P. Robert, and R. J. Ryder,
Approximate Bayesian computational methods.  Statistics and Compuing
22:1167--1180, 2012.

T. Toni and M. P. H. Stumpf, Simulation-based model selection for dynamical
systems in systems and population biology, Bioinformatics 26:104--110,
2010.

T. Toni, D. Welch, N. Strelkowa, A. Ipsen, and M. P. H. Stumpf, Approximate
Bayesian computation scheme for parameter inference and model selection in
dynamical systems Journal of the Royal Society, Interface 6:187--202, 2009.
}
\seealso{
\link[=proposals]{MCMC proposals}, and
the tutorials on the \href{https://kingaa.github.io/pomp/}{package website}.

Other summary statistics methods: \code{\link{basic_probes}},
  \code{\link{probe.match}}, \code{\link{probe}},
  \code{\link{spect}}
}
\author{
Edward L. Ionides, Aaron A. King
}
\concept{summary statistics methods}
