% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{accumulators}
\alias{accumulators}
\alias{accumvars}
\title{accumulators}
\description{
Accumulator variables
}
\details{
In formulating models, one sometimes wishes to define a state variable that will accumulate some quantity over the interval between successive observations.
\pkg{pomp} provides a facility to make such features more convenient.
Specifically, variables named in the \code{pomp}'s \code{accumvars} argument will be set to zero immediately following each observation.
See \code{\link{sir}} and the tutorials on the \href{https://kingaa.github.io/pomp/}{package website} for examples.
}
\seealso{
\code{\link{sir}}

Other information on model implementation: 
\code{\link{Csnippet}},
\code{\link{covariate_table}()},
\code{\link{distributions}},
\code{\link{dmeasure_spec}},
\code{\link{dprocess_spec}},
\code{\link{parameter_trans}()},
\code{\link{pomp-package}},
\code{\link{prior_spec}},
\code{\link{rinit_spec}},
\code{\link{rmeasure_spec}},
\code{\link{rprocess_spec}},
\code{\link{skeleton_spec}},
\code{\link{transformations}},
\code{\link{userdata}}
}
\concept{information on model implementation}
