% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pomp.R
\docType{methods}
\name{as_pomp}
\alias{as_pomp}
\title{as.pomp}
\usage{
as_pomp(object, ...)
}
\arguments{
\item{object}{the object to be coerced}

\item{\dots}{additional arguments}
}
\description{
Coerce to a \sQuote{pomp} object
}
\keyword{internal}
