\name{blowflies}
\alias{blowflies}
\alias{blowflies1}
\alias{blowflies2}
\docType{data}
\title{Model for Nicholson's blowflies.}
\description{
  \code{blowfly1} and \code{blowfly2} are \code{pomp} objects encoding stochastic delay-difference models.
}
\usage{data(blowflies)}
\examples{
data(blowflies)
plot(blowflies1)
plot(blowflies2)
}
\seealso{\code{\link{pomp-class}} and the vignettes}
\keyword{datasets}
