% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat}
\alias{concat}
\alias{concat,Pomp-method}
\alias{concat,Pfilter-method}
\alias{concat,Abc-method}
\alias{concat,Mif2-method}
\alias{concat,Pmcmc-method}
\title{Concatenate}
\usage{
\S4method{concat}{Pomp}(...)

\S4method{concat}{Pfilter}(...)

\S4method{concat}{Abc}(...)

\S4method{concat}{Mif2}(...)

\S4method{concat}{Pmcmc}(...)
}
\description{
Methods to concatenate objects into useful listie
}
\keyword{internal}
