% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_table.R
\docType{methods}
\name{Covariates}
\alias{Covariates}
\alias{covariate_table}
\alias{covariate_table,missing-method}
\alias{covariate_table,ANY-method}
\alias{covariate_table-ANY}
\title{Covariates}
\usage{
\S4method{covariate_table}{ANY}(..., times)
}
\arguments{
\item{\dots}{numeric vectors or data frames containing time-varying covariates}

\item{times}{the times corresponding to the covariates.
This may be given as a vector of (increasing, finite) numerical values.
Alternatively, one can indicate one of the variables given (either as a vector or as a data-frame column) by name or by index.}
}
\description{
Including time-varying covariates in a model.
}
