% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csnippet.R
\name{Csnippet}
\alias{Csnippet}
\alias{as,Csnippet-method}
\alias{Csnippet-class}
\alias{as-csnippet}
\alias{coerce,Csnippet,character-method}
\title{C snippets}
\usage{
Csnippet(text)
}
\description{
A class to hold snippets of C code.
}
\keyword{internal}
