\name{Csnippet}
\alias{Csnippet}
\alias{Csnippet-class}
\title{C code snippets}
\description{
  For including snippets of C code in \code{pomp} objects.
}
\usage{
  Csnippet(text);
}
\arguments{
  \item{text}{
    character; a snippt of C code.
  }
}
\value{
  An object of class \code{Csnippet}.
}
\author{Aaron A. King \email{kingaa at umich dot edu}}
\keyword{programming}
\keyword{interface}
