% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacca.R
\docType{data}
\name{dacca}
\alias{dacca}
\title{Model of cholera transmission for historic Bengal.}
\description{
\code{dacca} is a \sQuote{pomp} object containing census and cholera
mortality data from the Dacca district of the former British province of
Bengal over the years 1891 to 1940 together with a stochastic differential
equation transmission model.  The model is that of King et al. (2008).  The
parameters are the MLE for the SIRS model with seasonal reservoir.
}
\details{
Data are provided courtesy of Dr. Menno J. Bouma, London School of Tropical
Medicine and Hygiene.

\code{dacca} is a \sQuote{pomp} object containing the model, data, and MLE
parameters.  Parameters that naturally range over the positive reals are
log-transformed; parameters that range over the unit interval are
logit-transformed; parameters that are naturally unbounded or take integer
values are not transformed.
}
\examples{

pompExample(dacca)
plot(dacca)
## MLE:
coef(dacca)
plot(simulate(dacca))

}
\references{
A. A. King, E. L. Ionides, M. Pascual, and M. J. Bouma,
Inapparent infections and cholera dynamics, Nature, 454:877-880, 2008
}
\seealso{
Other pomp examples: \code{\link{blowflies}},
  \code{\link{gompertz}}, \code{\link{measles}},
  \code{\link{ou2}}, \code{\link{pompExample}},
  \code{\link{ricker}}, \code{\link{rw2}},
  \code{\link{sir_models}}

Other datasets: \code{\link{blowflies}},
  \code{\link{bsflu}}, \code{\link{measles}},
  \code{\link{parus}}
}
\concept{datasets}
\concept{pomp examples}
\keyword{datasets}
\keyword{models}
