% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_logLik.R, R/deprecated.R,
%   R/profileDesign.R, R/runifDesign.R, R/sliceDesign.R, R/sobolDesign.R
\name{cond.loglik}
\alias{cond.loglik}
\alias{deprecated}
\alias{profileDesign}
\alias{runifDesign}
\alias{sliceDesign}
\alias{sobolDesign}
\title{Deprecated functions.}
\usage{
cond.loglik(...)

profileDesign(
  ...,
  lower,
  upper,
  nprof,
  type = c("sobol", "runif"),
  stringsAsFactors = getOption("stringsAsFactors", FALSE)
)

runifDesign(lower = numeric(0), upper = numeric(0), nseq)

sliceDesign(center, ...)

sobolDesign(lower = numeric(0), upper = numeric(0), nseq)
}
\arguments{
\item{...}{In \code{profileDesign}, additional arguments specify the parameters over which to profile and the values of these parameters.
In \code{sliceDesign}, additional numeric vector arguments specify the locations of points along the slices.}

\item{lower, upper}{named numeric vectors giving the lower and upper bounds
of the ranges, respectively.}

\item{nprof}{The number of points per profile point.}

\item{type}{the type of design to use.
\code{type="sobol"} uses \code{\link{sobolDesign}};
\code{type="runif"} uses \code{\link{runifDesign}}.}

\item{stringsAsFactors}{should character vectors be converted to factors?}

\item{nseq}{Total number of points requested.}

\item{center}{\code{center} is a named numeric vector specifying the point
through which the slice(s) is (are) to be taken.}
}
\value{
\code{profileDesign} returns a data frame with \code{nprof} points per profile point.

\code{runifDesign} returns a data frame with \code{nseq} rows and one column for each variable named in \code{lower} and \code{upper}.

\code{sliceDesign} returns a data frame with one row per point.
The \sQuote{slice} variable indicates which slice the point belongs to.

\code{sobolDesign} returns a data frame with \code{nseq} rows and one column for each variable named in \code{lower} and \code{upper}.
}
\description{
These functions are obsolete and will be removed in a future release.

Obsolete: replaced by \code{\link{profile_design}}.

Obsolete: replaced by \code{runif_design}.

Obsolete: replaced by \code{slice_design}.

Obsolete: replaced by \code{sobol_design}.
}
\details{
The Sobol' sequence generation is performed using codes from the
\href{https://nlopt.readthedocs.io/en/latest/}{\pkg{NLopt} library} by S. Johnson.
}
\author{
Aaron A. King
}
\keyword{internal}
