% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{conv.rec}
\alias{values}
\alias{onestep.dens}
\alias{onestep.sim}
\alias{discrete.time.sim}
\alias{euler.sim}
\alias{gillespie.sim}
\alias{gillespie.hl.sim}
\title{Deprecated functions}
\usage{
conv.rec(object, ...)

values(object, ...)

onestep.dens(dens.fun, PACKAGE)

onestep.sim(step.fun)

discrete.time.sim(step.fun, delta.t = 1)

euler.sim(step.fun, delta.t)

gillespie.sim(rate.fun, v, hmax = Inf)

gillespie.hl.sim(..., .pre = "", .post = "", hmax = Inf)
}
\description{
Deprecated functions
}
\keyword{internal}
