% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Deprecated functions}
\alias{Deprecated functions}
\alias{conv.rec}
\alias{values}
\alias{onestep.dens}
\title{Deprecated functions}
\usage{
conv.rec(object, ...)

values(object, ...)

onestep.dens(dens.fun, PACKAGE)
}
\description{
Deprecated functions
}
\keyword{internal}
