% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinit_spec.R
\name{dinit specification}
\alias{dinit specification}
\title{The initial-state distribution}
\description{
Specification of the initial-state distribution density evaluator, dinit.
}
\details{
To fully specify the unobserved Markov state process, one must give its distribution at the zero-time (\code{t0}).
One specifies how to evaluate the log probability density function for this distribution using the \code{dinit} argument.
As usual, this can be provided either as a C snippet or as an \R function.
In the former case, bear in mind that:
\enumerate{
  \item The goal of a this snippet is computation of a log likelihood, to be put into a variable named \code{loglik}.
  \item In addition to the state variables, parameters, and covariates (if any), the variable \code{t}, containing the zero-time, will be defined in the context in which the snippet is executed.
}
\link[=Csnippet]{General rules for writing C snippets can be found here}.

If an \R function is to be used, pass
\preformatted{
   dinit = f
}
to \code{pomp}, where \code{f} is a function with arguments that can include the time \code{t}, any or all of the model state variables, parameters, and covariates.
As usual, \code{f} may take additional arguments, provided these are passed along with it in the call to \code{pomp}.
\code{f} must return a single numeric value, the log likelihood.
}
\section{Note for Windows users}{


Some Windows users report problems when using C snippets in parallel computations.
These appear to arise when the temporary files created during the C snippet compilation process are not handled properly by the operating system.
To circumvent this problem, use the \code{\link[=pomp]{cdir}} and \code{\link[=pomp]{cfile}} options to cause the C snippets to be written to a file of your choice, thus avoiding the use of temporary files altogether.
}

\seealso{
\code{\link{dinit}}

More on implementing POMP models: 
\code{\link{Csnippet}},
\code{\link{accumulator variables}},
\code{\link{basic components}},
\code{\link{betabinomial}},
\code{\link{covariates}},
\code{\link{distributions}},
\code{\link{dmeasure specification}},
\code{\link{dprocess specification}},
\code{\link{emeasure specification}},
\code{\link{parameter transformations}},
\code{\link{pomp-package}},
\code{\link{pomp}},
\code{\link{prior specification}},
\code{\link{rinit specification}},
\code{\link{rmeasure specification}},
\code{\link{rprocess specification}},
\code{\link{skeleton specification}},
\code{\link{transformations}},
\code{\link{userdata}},
\code{\link{vmeasure specification}}
}
\concept{implementation information}
