% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\docType{methods}
\name{dprior}
\alias{dprior}
\alias{dprior,pomp-method}
\alias{dprior-pomp}
\title{dprior}
\usage{
\S4method{dprior}{pomp}(object, params, log = FALSE, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}.}

\item{params}{a \code{npar} x \code{nrep} matrix of parameters.
Each column is an independent parameter set and is paired with the corresponding column of \code{x} or \code{xstart}.}

\item{log}{if TRUE, log probabilities are returned.}

\item{...}{additional arguments are ignored.}
}
\value{
The required density (or log density), as a numeric vector.
}
\description{
Evaluates the prior probability density.
}
\seealso{
Other pomp workhorses: \code{\link{dmeasure}},
  \code{\link{dprocess}}, \code{\link{partrans}},
  \code{\link{rinit}}, \code{\link{rmeasure}},
  \code{\link{rprior}}, \code{\link{rprocess}},
  \code{\link{skeleton}}, \code{\link{workhorses}}
}
\concept{pomp workhorses}
\keyword{internal}
