% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\docType{methods}
\name{dprocess}
\alias{dprocess}
\alias{dprocess,pomp-method}
\alias{dprocess-pomp}
\title{dprocess}
\usage{
\S4method{dprocess}{pomp}(object, x, times, params, log = FALSE, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}.}

\item{x}{a rank-3 array containing states of the unobserved process.
The dimensions of \code{x} are \code{nvars} x \code{nrep} x \code{ntimes},
where \code{nvars} is the number of state variables,
\code{nrep} is the number of replicates,
and \code{ntimes} is the length of \code{times}.}

\item{times}{a numeric vector (length \code{ntimes}) containing times.
These must be in non-decreasing order.}

\item{params}{a \code{npar} x \code{nrep} matrix of parameters.
Each column is an independent parameter set and is paired with the corresponding column of \code{x} or \code{xstart}.}

\item{log}{if TRUE, log probabilities are returned.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{dprocess} returns a matrix of dimensions \code{nrep} x \code{ntimes-1}.
If \code{d} is the returned matrix, \code{d[j,k]} is the likelihood (or the log likelihood if \code{log=TRUE}) of the transition from state \code{x[,j,k-1]} at time \code{times[k-1]} to state \code{x[,j,k]} at time \code{times[k]}.
}
\description{
Evaluates the probability density of a sequence of consecutive state transitions.
}
\seealso{
Other pomp workhorses: \code{\link{dmeasure}},
  \code{\link{dprior}}, \code{\link{partrans}},
  \code{\link{rinit}}, \code{\link{rmeasure}},
  \code{\link{rprior}}, \code{\link{rprocess}},
  \code{\link{skeleton}}, \code{\link{workhorses}}
}
\concept{pomp workhorses}
\keyword{internal}
