% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola.R
\docType{data}
\name{ebola}
\alias{ebola}
\alias{ebolaModel}
\alias{ebolaWHO}
\title{Ebola outbreak, West Africa}
\usage{
ebolaModel(country = c("Guinea", "SierraLeone", "Liberia", "WestAfrica"),
  data = NULL, timestep = 0.01, nstageE = 3L, R0 = 1.4,
  rho = 0.2, cfr = 0.7, k = 0, incubation_period = 11.4/7,
  infectious_period = 7/7, type = c("raw", "cum"), na.rm = FALSE,
  least.sq = FALSE)
}
\arguments{
\item{country}{name of country or \dQuote{WestAfrica} for the regional outbreak.}

\item{data}{if NULL, the situation report data (see \code{ebolaWHO}) for the appropriate country or region will be used.
Providing a dataset here will override this behavior.}

\item{timestep}{duration (in weeks) of Euler time-step for the simulations.}

\item{nstageE}{integer; number of incubation stages.}

\item{incubation_period, infectious_period}{mean duration (in weeks) of the incubation and infectious periods.}

\item{type}{should raw or cumulative case counts be used as the data?}

\item{na.rm}{should missing data be removed?}

\item{least.sq}{logical; if \code{TRUE}, ordinary least squares will be used.}
}
\description{
Data and models for the 2013--2015 outbreak of Ebola virus disease in West Africa.
}
\details{
The data include case counts according to the WHO situation report dated 1 October 2014.
The models are described in King et al. (2015).
}
\section{Model structure}{

The observation model is a hierarchical model for cases and deaths:
  \deqn{p(R_t, D_t| C_t) = p(R_t | C_t) * p(D_t | C_t, R_t).}
Here, \eqn{p(R_t | C_t)} is negative binomial with mean \eqn{\rho C_t} and dispersion parameter \eqn{1/k};
\eqn{p(D_t | C_t, R_t)} is binomial with size \eqn{R_t} and probability equal to the case fatality rate \code{cfr}.

The default incubation period is supposed to be Gamma distributed with shape parameter `nstageE` and mean 11.4 days.
The discrete-time formula is used to calculate the corresponding `alpha`` (cf. He et al. 2010).
The case-fatality ratio (`cfr`) is fixed at 0.7 (cf. WHO Ebola Response Team 2014).
}

\references{
A. A. King, M. Domenech de Cellès, F. M. G. Magpantay, and P. Rohani.
Avoidable errors in the modelling of outbreaks of emerging pathogens, with special reference to Ebola.
Proceedings of the Royal Society of London, Series B, 282: 20150347, 2015.

WHO Ebola Response Team.
Ebola virus disease in West Africa---the first 9 months of the epidemic and forward projections.
New England Journal of Medicine 371: 1481--1495, 2014.

D. He, E. L. Ionides, & A. A. King.
Plug-and-play inference for disease dynamics:
measles in large and small populations as a case study.
J. R. Soc. Interface, 7: 271--283, 2010.
}
\seealso{
Other datasets: \code{\link{blowflies}},
  \code{\link{bsflu}}, \code{\link{dacca}},
  \code{\link{measles}}, \code{\link{parus}}

Other pomp examples: \code{\link{blowflies}},
  \code{\link{dacca}}, \code{\link{gompertz}},
  \code{\link{measles}}, \code{\link{ou2}},
  \code{\link{ricker}}, \code{\link{rw2}},
  \code{\link{sir_models}}, \code{\link{verhulst}}
}
\concept{datasets}
\concept{pomp examples}
