% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_sample_size.R
\docType{methods}
\name{Effective sample size}
\alias{Effective sample size}
\alias{eff.sample.size}
\alias{eff.sample.size,missing-method}
\alias{eff.sample.size,ANY-method}
\alias{eff.sample.size-bsmcd_pomp}
\alias{eff.sample.size,bsmcd_pomp-method}
\alias{eff.sample.size-pfilterd_pomp}
\alias{eff.sample.size,pfilterd_pomp-method}
\title{Effective sample size}
\usage{
eff.sample.size(object, ...)

\S4method{eff.sample.size}{bsmcd_pomp}(object, ...)

\S4method{eff.sample.size}{pfilterd_pomp}(object, ...)
}
\arguments{
\item{object}{object extending class \sQuote{pomp}}

\item{\dots}{ignored}
}
\description{
Estimate the effective sample size of a Monte Carlo computation.
}
