% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_algorithms.R
\name{estimation algorithms}
\alias{estimation algorithms}
\title{Parameter estimation algorithms for POMP models.}
\description{
\pkg{pomp} currently implements the following algorithms for estimating model parameters:
\itemize{
\item \link[=mif2]{iterated filtering (IF2)}
\item \link[=pmcmc]{particle Markov chain Monte Carlo (PMCMC)}
\item \link[=abc]{approximate Bayesian computation (ABC)}
\item \link[=probe matching]{probe-matching via synthetic likelihood}
\item \link{nonlinear forecasting}
\item \link[=spectrum matching]{power-spectrum matching}
\item \link[=bsmc2]{Liu-West Bayesian sequential Monte Carlo}
\item \link[=kalman]{Ensemble and ensemble-adjusted Kalman filters}
}
}
\details{
Help pages detailing each estimation algorithm are provided.
}
\seealso{
\link[=basic components]{basic model components},
\link[=workhorses]{workhorse functions},
\link[=elementary algorithms]{elementary algorithms}.

More on \pkg{pomp} estimation algorithms:
\code{\link{approximate Bayesian computation}},
\code{\link{bsmc2}()},
\code{\link{mif2}()},
\code{\link{nonlinear forecasting}},
\code{\link{pmcmc}()},
\code{\link{pomp-package}},
\code{\link{probe matching}},
\code{\link{spectrum matching}}
}
\concept{estimation methods}
