% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mean.R
\docType{methods}
\name{filter.mean}
\alias{filter.mean}
\alias{filter.mean,ANY-method}
\alias{filter.mean,missing-method}
\alias{filter.mean-kalmand_pomp}
\alias{filter.mean,kalmand_pomp-method}
\alias{filter.mean-pfilterd_pomp}
\alias{filter.mean,pfilterd_pomp-method}
\title{Filtering mean}
\usage{
\S4method{filter.mean}{kalmand_pomp}(object, vars, ...)

\S4method{filter.mean}{pfilterd_pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{optional character; names of variables}

\item{\dots}{ignored}
}
\description{
The mean of the filtering distribution
}
\details{
The filtering distribution is that of
\deqn{X_t \vert y_1,\dots,y_t,}{X_t | y_1,\dots,y_t,}
where \eqn{X_t}, \eqn{y_t} are the state vector and data, respectively,
at time \eqn{t}.
}
\seealso{
Other particle filter methods: \code{\link{bsmc2}},
  \code{\link{cond.logLik}}, \code{\link{eff.sample.size}},
  \code{\link{mif2}}, \code{\link{pfilter}},
  \code{\link{pmcmc}}, \code{\link{pred.mean}},
  \code{\link{pred.var}}
}
\concept{particle filter methods}
