% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mean.R
\docType{methods}
\name{Filtering mean}
\alias{Filtering mean}
\alias{filter.mean}
\alias{filter.mean,ANY-method}
\alias{filter.mean,missing-method}
\alias{filter.mean-kalmand_pomp}
\alias{filter.mean,kalmand_pomp-method}
\alias{filter.mean-pfilterd_pomp}
\alias{filter.mean,pfilterd_pomp-method}
\title{Filtering mean}
\usage{
\S4method{filter.mean}{kalmand_pomp}(object, vars, ...)

\S4method{filter.mean}{pfilterd_pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{optional character; names of variables}

\item{\dots}{ignored}
}
\description{
The mean of the filtering distribution
}
