% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_traj.R
\docType{methods}
\name{Filtering trajectories}
\alias{Filtering trajectories}
\alias{filter.traj}
\alias{filter.traj,ANY-method}
\alias{filter.traj,missing-method}
\alias{filter.traj-pfilterd_pomp}
\alias{filter.traj,pfilterd_pomp-method}
\alias{filter.traj-pfilterList}
\alias{filter.traj,pfilterList-method}
\alias{filter.traj-pmcmcd_pomp}
\alias{filter.traj,pmcmcd_pomp-method}
\alias{filter.traj-pmcmcList}
\alias{filter.traj,pmcmcList-method}
\title{Filtering trajectories}
\usage{
\S4method{filter.traj}{pfilterd_pomp}(object, vars, ...)

\S4method{filter.traj}{pfilterList}(object, ...)

\S4method{filter.traj}{pmcmcd_pomp}(object, vars, ...)

\S4method{filter.traj}{pmcmcList}(object, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{optional character; names of variables}

\item{...}{ignored}
}
\description{
Trajectories drawn from the smoothing distribution
}
