% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\docType{methods}
\name{flow}
\alias{flow}
\alias{flow,missing-method}
\alias{flow,ANY-method}
\alias{flow-pomp}
\alias{flow,pomp-method}
\title{Flow of a deterministic model}
\usage{
\S4method{flow}{pomp}(object, xstart, params, times, offset = 0L, ...,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}.
This will typically be the output of \code{pomp}, \code{simulate}, or one of the \pkg{pomp} inference algorithms.}

\item{xstart}{an array with dimensions \code{nvar} x \code{nrep} giving the initial conditions of the trajectories to be computed.}

\item{params}{a \code{npar} x \code{nrep} matrix of parameters.
Each column is treated as an independent parameter set, in correspondence with the corresponding column of \code{x}.}

\item{times}{a numeric vector (length \code{ntimes}) containing times.
These must be in strictly increasing order.
The first entry is assumed to be the time of \code{xstart.}}

\item{offset}{a non-negative integer scalar.
By default, \code{offset = 0}.
See below (Value) for the effect of \code{offset}.}

\item{...}{Additional arguments are passed to the ODE integrator (if the skeleton is a vectorfield) and are ignored if it is a map.
See \code{\link[deSolve]{ode}} for a description of the additional arguments accepted by the ODE integrator.}

\item{verbose}{logical; if \code{TRUE}, diagnostic messages will be printed to the console.}
}
\value{
\code{flow} returns an array of dimensions \code{nvar} x \code{nrep} x \code{ntimes-offset}.
If \code{x} is the returned matrix, \code{x[i,j,k]} is the i-th component of the state vector at time \code{times[k+offset]} given parameters \code{params[,j]}.
}
\description{
Compute the flow induced by a deterministic dynamical system, whether vectorfield or map.
}
\details{
In the case of a discrete-time system (map), \code{flow} iterates the map to yield trajectories of the system.
In the case of a continuous-time system (vectorfield), \code{flow} uses the numerical solvers in \pkg{\link[deSolve]{deSolve}} to integrate the vectorfield starting from given initial conditions.
}
\seealso{
\code{\link{skeleton}}, \code{\link{trajectory}}, \code{\link{rprocess}}

Other pomp workhorses: \code{\link{dmeasure}},
  \code{\link{dprior}}, \code{\link{dprocess}},
  \code{\link{partrans}}, \code{\link{rinit}},
  \code{\link{rmeasure}}, \code{\link{rprior}},
  \code{\link{rprocess}}, \code{\link{skeleton}},
  \code{\link{workhorses}}
}
\concept{pomp workhorses}
