% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\alias{forecast,missing-method}
\alias{forecast,ANY-method}
\alias{forecast,kalmand_pomp-method}
\title{Forecast mean}
\usage{
forecast(object, ...)

\S4method{forecast}{kalmand_pomp}(object, vars, ...)
}
\arguments{
\item{object}{result of a filtering computation}

\item{...}{ignored}

\item{vars}{optional character; names of variables}
}
\description{
Mean of the one-step-ahead forecasting distribution.
}
\seealso{
Other extraction methods: 
\code{\link{coef}()},
\code{\link{cond.logLik}()},
\code{\link{covmat}()},
\code{\link{eff.sample.size}()},
\code{\link{filter.mean}()},
\code{\link{filter.traj}()},
\code{\link{logLik}},
\code{\link{obs}()},
\code{\link{pred.mean}()},
\code{\link{pred.var}()},
\code{\link{saved.states}()},
\code{\link{spy}()},
\code{\link{states}()},
\code{\link{summary}()},
\code{\link{timezero}()},
\code{\link{time}()},
\code{\link{traces}()}
}
\concept{extraction methods}
