% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\alias{forecast,missing-method}
\alias{forecast,ANY-method}
\alias{forecast,kalmand_pomp-method}
\alias{forecast,pfilterd_pomp-method}
\title{Forecast mean}
\usage{
forecast(object, ...)

\S4method{forecast}{kalmand_pomp}(object, vars, ..., format = c("array", "data.frame"))

\S4method{forecast}{pfilterd_pomp}(object, vars, ..., format = c("array", "data.frame"))
}
\arguments{
\item{object}{result of a filtering computation}

\item{...}{ignored}

\item{vars}{optional character; names of variables}

\item{format}{format of the returned object.}
}
\description{
Mean of the one-step-ahead forecasting distribution.
}
\seealso{
Other extraction methods: 
\code{\link{coef}()},
\code{\link{cond_logLik}()},
\code{\link{covmat}()},
\code{\link{eff_sample_size}()},
\code{\link{filter_mean}()},
\code{\link{filter_traj}()},
\code{\link{logLik}},
\code{\link{obs}()},
\code{\link{pred_mean}()},
\code{\link{pred_var}()},
\code{\link{saved_states}()},
\code{\link{spy}()},
\code{\link{states}()},
\code{\link{summary}()},
\code{\link{timezero}()},
\code{\link{time}()},
\code{\link{traces}()}
}
\concept{extraction methods}
