% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\docType{methods}
\name{Forecast}
\alias{Forecast}
\alias{forecast}
\alias{forecast,missing-method}
\alias{forecast,ANY-method}
\alias{forecast-kalmand_pomp}
\alias{forecast,kalmand_pomp-method}
\title{Forecast mean}
\usage{
forecast(object, ...)

\S4method{forecast}{kalmand_pomp}(object, vars, ...)
}
\arguments{
\item{object}{result of a filtering computation}

\item{...}{ignored}

\item{vars}{optional character; names of variables}
}
\description{
Mean of the one-step-ahead forecasting distribution.
}
