% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman.R
\docType{methods}
\name{Ensemble Kalman filter methods}
\alias{Ensemble Kalman filter methods}
\alias{enkf}
\alias{eakf}
\alias{enkf,ANY-method}
\alias{enkf,missing-method}
\alias{eakf,ANY-method}
\alias{eakf,missing-method}
\alias{enkf-pomp}
\alias{enkf,pomp-method}
\alias{eakf-pomp}
\alias{eakf,pomp-method}
\title{Ensemble Kalman filters}
\usage{
\S4method{enkf}{pomp}(object, params, Np, h, R, ...,
  verbose = getOption("verbose", FALSE))

\S4method{eakf}{pomp}(object, params, Np, C, R, ...,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{object}{An object of class \sQuote{pomp} or extending class
\sQuote{pomp}.}

\item{params}{optional named numeric vector containing the parameters at
which the filtering should be performed.  By default, \code{params =
coef(object)}.}

\item{Np}{the number of particles to use.}

\item{h}{function returning the expected value of the observation given the
state.}

\item{R}{matrix; variance of the measurement noise.}

\item{\dots}{additional arguments, which can be used to modify components of the model.}

\item{verbose}{logical; if \code{TRUE}, progress information is reported.}

\item{C}{matrix converting state vector into expected value of the
observation.}
}
\value{
An object of class \sQuote{kalmand_pomp}.
}
\description{
The ensemble Kalman filter and ensemble adjustment Kalman filter.
}
\references{
Evensen, G. (1994) Sequential data assimilation with a
nonlinear quasi-geostrophic model using Monte Carlo methods to forecast
error statistics Journal of Geophysical Research: Oceans 99:10143--10162

Evensen, G. (2009) Data assimilation: the ensemble Kalman filter
Springer-Verlag.

Anderson, J. L. (2001) An Ensemble Adjustment Kalman Filter for Data
Assimilation Monthly Weather Review 129:2884--2903
}
\seealso{
\code{\link{pfilter}}, and the tutorials on
the \href{https://kingaa.github.io/pomp}{package website}.
}
\author{
Aaron A. King
}
