% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman.R
\docType{methods}
\name{kalman}
\alias{kalman}
\alias{enkf}
\alias{eakf}
\alias{enkf,ANY-method}
\alias{enkf,missing-method}
\alias{eakf,ANY-method}
\alias{eakf,missing-method}
\alias{enkf-data.frame}
\alias{enkf,data.frame-method}
\alias{enkf-pomp}
\alias{enkf,pomp-method}
\alias{eakf-data.frame}
\alias{eakf,data.frame-method}
\alias{eakf-pomp}
\alias{eakf,pomp-method}
\title{Ensemble Kalman filters}
\usage{
\S4method{enkf}{data.frame}(data, rinit, rprocess, params, Np, h, R, ...,
  verbose = getOption("verbose", FALSE))

\S4method{enkf}{pomp}(data, Np, h, R, ..., verbose = getOption("verbose",
  FALSE))

\S4method{eakf}{data.frame}(data, rinit, rprocess, params, Np, C, R, ...,
  verbose = getOption("verbose", FALSE))

\S4method{eakf}{pomp}(data, Np, C, R, ..., verbose = getOption("verbose",
  FALSE))
}
\arguments{
\item{data}{either a data frame holding the time series data, or an object of class \sQuote{pomp}, i.e., the output of one of \pkg{pomp}'s methods.}

\item{rinit}{simulator of the initial-state distribution.
This can be furnished either as a C snippet, an \R function, or the name of a pre-compiled native routine available in a dynamically loaded library.
For more information on specifying \code{rinit}, see \link[=rinit_plugins]{here}.}

\item{rprocess}{simulator of the latent state process, specified either as a C snippet, an \R function, or the name of a pre-compiled native routine available in a dynamically loaded library.
For more information on specifying \code{rprocess}, see \link[=rprocess_plugins]{here}.}

\item{params}{optional; named numeric vector of parameters.  This will be
coerced internally to storage mode \code{double}.}

\item{Np}{the number of particles to use.}

\item{h}{function returning the expected value of the observation given the
state.}

\item{R}{matrix; variance of the measurement noise.}

\item{...}{additional arguments supply new or modify existing model characteristics or components.}

\item{verbose}{if \code{TRUE}, diagnostic messages will be printed to the console}

\item{C}{matrix converting state vector into expected value of the
observation.}
}
\value{
An object of class \sQuote{kalmand_pomp}.
}
\description{
The ensemble Kalman filter and ensemble adjustment Kalman filter.
}
\references{
Evensen, G. (1994) Sequential data assimilation with a
nonlinear quasi-geostrophic model using Monte Carlo methods to forecast
error statistics Journal of Geophysical Research: Oceans 99:10143--10162

Evensen, G. (2009) Data assimilation: the ensemble Kalman filter
Springer-Verlag.

Anderson, J. L. (2001) An Ensemble Adjustment Kalman Filter for Data
Assimilation Monthly Weather Review 129:2884--2903
}
\seealso{
\code{\link{pfilter}}, and the tutorials on the \href{https://kingaa.github.io/pomp}{package website}.
}
\author{
Aaron A. King
}
