% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R, R/loglik.R, R/probe_match.R,
%   R/spect_match.R, R/traj_match.R
\docType{methods}
\name{logLik-probed_pomp}
\alias{logLik-probed_pomp}
\alias{logLik,probed_pomp-method}
\alias{logLik}
\alias{logLik,ANY-method}
\alias{logLik,missing-method}
\alias{logLik,listies-method}
\alias{logLik-generic}
\alias{logLik-pfilterd_pomp}
\alias{logLik,pfilterd_pomp-method}
\alias{logLik-kalmand_pomp}
\alias{logLik,kalmand_pomp-method}
\alias{logLik-pmcmcd_pomp}
\alias{logLik,pmcmcd_pomp-method}
\alias{logLik-nlfd_pomp}
\alias{logLik,nlfd_pomp-method}
\alias{logLik-bsmcd_pomp}
\alias{logLik,bsmcd_pomp-method}
\alias{logLik-probe_match_obfjun}
\alias{logLik,probe_match_objfun-method}
\alias{logLik-spect_match_objfun}
\alias{logLik,spect_match_objfun-method}
\alias{logLik-traj_match_objfun}
\alias{logLik,traj_match_objfun-method}
\title{Log likelihood}
\usage{
\S4method{logLik}{probed_pomp}(object)

logLik(object, ...)

\S4method{logLik}{pfilterd_pomp}(object)

\S4method{logLik}{kalmand_pomp}(object)

\S4method{logLik}{pmcmcd_pomp}(object)

\S4method{logLik}{nlfd_pomp}(object, ...)

\S4method{logLik}{bsmcd_pomp}(object)

\S4method{logLik}{probe_match_objfun}(object)

\S4method{logLik}{spect_match_objfun}(object)

\S4method{logLik}{traj_match_objfun}(object)
}
\arguments{
\item{object}{fitted model object}

\item{\dots}{ignored}
}
\value{
When \code{object} is of \sQuote{probed_pomp} class (i.e., the result of a \code{probe} computation), \code{logLik} retrieves the \dQuote{synthetic likelihood} (see \code{\link{probe}}).

numerical value of the log likelihood.
Note that some methods compute not the log likelihood itself but instead a related quantity.
To keep the code simple, the \code{logLik} function is nevertheless used to extract this quantity.

When \code{object} is of \sQuote{nlfd_pomp} class (i.e., the result of a \code{nlf} computation), \code{logLik} retrieves the \dQuote{simulate maximum quasi log likelihood} (see \code{\link{nlf}}).

When \code{object} is of \sQuote{bsmcd_pomp} class (i.e., the result of a \code{bsmc2} computation), \code{logLik} retrieves the \dQuote{log evidence} (see \code{\link{bsmc2}}).
}
\description{
Extract the estimated log likelihood from a fitted model.
}
