% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup}
\alias{lookup}
\title{Lookup table}
\usage{
lookup(table, t)
}
\arguments{
\item{table}{a \sQuote{covartable} object created by a call to \code{\link{covariate_table}}}

\item{t}{numeric vector; times at which interpolated values of the covariates in \code{table} are required.}
}
\value{
A numeric vector or matrix of the interpolated values.
}
\description{
Interpolate values from a lookup table
}
\details{
A warning will be generated if extrapolation is performed.
}
\seealso{
Other interpolation: 
\code{\link{bsplines}},
\code{\link{covariate_table}()}
}
\concept{interpolation}
