% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{melt}
\alias{melt,pomp-method}
\title{Objects exported from other packages}
\usage{
\S4method{melt}{pomp}(data, ..., na.rm = FALSE, value.name = "value")
}
\arguments{
\item{data}{Data set to melt}

\item{...}{further arguments passed to or from other methods.}

\item{na.rm}{Should NA values be removed from the data set? This will
convert explicit missings to implicit missings.}

\item{value.name}{name of variable used to store values}
}
\details{
See \code{reshape2::\link[reshape2]{melt}} for details.

A \sQuote{pomp} object can be melted into a data frame.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reshape2}{\code{\link[reshape2]{melt}}}
}}

