% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\docType{methods}
\name{obs}
\alias{obs}
\alias{obs,pomp-method}
\title{obs}
\usage{
\S4method{obs}{pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{names of variables to retrieve}

\item{\dots}{ignored}
}
\description{
Extract the data array from a \sQuote{pomp} object.
}
