% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\docType{methods}
\name{obs}
\alias{obs}
\alias{obs,pomp-method}
\title{obs}
\usage{
\S4method{obs}{pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{names of variables to retrieve}

\item{\dots}{ignored}
}
\description{
Extract the data array from a \sQuote{pomp} object.
}
\seealso{
Other extraction methods: 
\code{\link{coef}()},
\code{\link{cond.logLik}()},
\code{\link{covmat}()},
\code{\link{eff.sample.size}()},
\code{\link{filter.mean}()},
\code{\link{filter.traj}()},
\code{\link{forecast}()},
\code{\link{logLik}},
\code{\link{pred.mean}()},
\code{\link{pred.var}()},
\code{\link{saved.states}()},
\code{\link{spy}()},
\code{\link{states}()},
\code{\link{summary}()},
\code{\link{timezero}()},
\code{\link{time}()},
\code{\link{traces}()}
}
\concept{extraction methods}
