% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\docType{methods}
\name{obs}
\alias{obs}
\alias{obs,pomp-method}
\alias{obs,listie-method}
\title{obs}
\usage{
\S4method{obs}{pomp}(object, vars, ...)

\S4method{obs}{listie}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{names of variables to retrieve}

\item{\dots}{ignored}
}
\description{
Extract the data array from a \sQuote{pomp} object.
}
\seealso{
Other extraction methods: 
\code{\link{coef}()},
\code{\link{cond_logLik}()},
\code{\link{covmat}()},
\code{\link{eff_sample_size}()},
\code{\link{filter_mean}()},
\code{\link{filter_traj}()},
\code{\link{forecast}()},
\code{\link{logLik}},
\code{\link{pred_mean}()},
\code{\link{pred_var}()},
\code{\link{saved_states}()},
\code{\link{spy}()},
\code{\link{states}()},
\code{\link{summary}()},
\code{\link{timezero}()},
\code{\link{time}()},
\code{\link{traces}()}
}
\concept{extraction methods}
