% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_trans.R
\docType{methods}
\name{Parameter transformations}
\alias{Parameter transformations}
\alias{parameter_trans}
\alias{parameter_trans,missing,missing-method}
\alias{parameter_trans,ANY,ANY-method}
\alias{parameter_trans,ANY,missing-method}
\alias{parameter_trans,NULL,NULL-method}
\alias{parameter_trans,function,function-method}
\alias{parameter_trans,missing,ANY-method}
\alias{parameter_trans,pomp_fun,pomp_fun-method}
\alias{parameter_trans-missing,missing}
\alias{parameter_trans-Csnippet,Csnippet}
\alias{parameter_trans,Csnippet,Csnippet-method}
\alias{parameter_trans-character,character}
\alias{parameter_trans,character,character-method}
\alias{parameter_trans-function,function}
\title{Parameter transformations}
\usage{
\S4method{parameter_trans}{missing,missing}(..., log, logit, barycentric)

\S4method{parameter_trans}{Csnippet,Csnippet}(toEst, fromEst, ..., log,
  logit, barycentric)

\S4method{parameter_trans}{character,character}(toEst, fromEst, ...)

\S4method{parameter_trans}{`function`,`function`}(toEst, fromEst, ...)
}
\description{
Equipping models with parameter transformations.
}
\keyword{internal}
