% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\docType{methods}
\name{partrans}
\alias{partrans}
\alias{partrans,pomp-method}
\alias{partrans-pomp}
\alias{partrans,ANY-method}
\alias{partrans,missing-method}
\title{partrans}
\usage{
\S4method{partrans}{pomp}(object, params, dir = c("fromEst", "toEst"),
  ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}.
This will typically be the output of \code{pomp}, \code{simulate}, or one of the \pkg{pomp} inference algorithms.}

\item{params}{a \code{npar} x \code{nrep} matrix of parameters.
Each column is treated as an independent parameter set, in correspondence with the corresponding column of \code{x}.}

\item{dir}{the direction of the transformation to perform.}

\item{...}{additional arguments are ignored.}
}
\value{
If \code{dir=fromEst}, the parameters in \code{params} are assumed to be on the estimation scale and are transformed onto the natural scale.
If \code{dir=toEst}, they are transformed onto the estimation scale.
In both cases, the parameters are returned as a named numeric vector or an array with rownames, as appropriate.
}
\description{
Performs parameter transformations.
}
\seealso{
Specification of parameter transformations: \code{\link{parameter_trans}}

Other pomp workhorses: \code{\link{dmeasure}},
  \code{\link{dprior}}, \code{\link{dprocess}},
  \code{\link{rinit}}, \code{\link{rmeasure}},
  \code{\link{rprior}}, \code{\link{rprocess}},
  \code{\link{skeleton}}, \code{\link{workhorses}}
}
\concept{pomp workhorses}
\keyword{internal}
