% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_mean.R
\docType{methods}
\name{Prediction mean}
\alias{Prediction mean}
\alias{pred.mean}
\alias{pred.mean,ANY-method}
\alias{pred.mean,missing-method}
\alias{pred.mean-kalmand_pomp}
\alias{pred.mean,kalmand_pomp-method}
\alias{pred.mean-pfilterd_pomp}
\alias{pred.mean,pfilterd_pomp-method}
\title{Prediction mean}
\usage{
\S4method{pred.mean}{kalmand_pomp}(object, vars, ...)

\S4method{pred.mean}{pfilterd_pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{optional character; names of variables}

\item{...}{ignored}
}
\description{
The mean of the prediction distribution
}
