% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_mean.R
\docType{methods}
\name{pred.mean}
\alias{pred.mean}
\alias{pred.mean,ANY-method}
\alias{pred.mean,missing-method}
\alias{pred.mean-kalmand_pomp}
\alias{pred.mean,kalmand_pomp-method}
\alias{pred.mean-pfilterd_pomp}
\alias{pred.mean,pfilterd_pomp-method}
\title{Prediction mean}
\usage{
\S4method{pred.mean}{kalmand_pomp}(object, vars, ...)

\S4method{pred.mean}{pfilterd_pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{optional character; names of variables}

\item{...}{ignored}
}
\description{
The mean of the prediction distribution
}
\details{
The prediction distribution is that of
\deqn{X_t \vert y_1,\dots,y_{t-1},}{X_t | y_1,\dots,y_(t-1),}
where \eqn{X_t}, \eqn{y_t} are the state vector and data, respectively,
at time \eqn{t}.
}
\seealso{
Other particle filter methods: \code{\link{cond.logLik}},
  \code{\link{eff.sample.size}}, \code{\link{filter.mean}},
  \code{\link{mif2}}, \code{\link{pfilter}},
  \code{\link{pmcmc}}, \code{\link{pred.var}}
}
\concept{particle filter methods}
