% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_var.R
\docType{methods}
\name{Prediction variance}
\alias{Prediction variance}
\alias{pred.var}
\alias{pred.var,ANY-method}
\alias{pred.var,missing-method}
\alias{pred.var-pfilterd_pomp}
\alias{pred.var,pfilterd_pomp-method}
\title{Prediction variance}
\usage{
\S4method{pred.var}{pfilterd_pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{optional character; names of variables}

\item{...}{ignored}
}
\description{
The variance of the prediction distribution
}
