% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_var.R
\docType{methods}
\name{pred.var}
\alias{pred.var}
\alias{pred.var,ANY-method}
\alias{pred.var,missing-method}
\alias{pred.var-pfilterd_pomp}
\alias{pred.var,pfilterd_pomp-method}
\title{Prediction variance}
\usage{
\S4method{pred.var}{pfilterd_pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{optional character; names of variables}

\item{...}{ignored}
}
\description{
The variance of the prediction distribution
}
\details{
The prediction distribution is that of
\deqn{X_t \vert y_1,\dots,y_{t-1},}{X_t | y_1,\dots,y_(t-1),}
where \eqn{X_t}, \eqn{y_t} are the state vector and data, respectively,
at time \eqn{t}.
}
\seealso{
Other particle filter methods: \code{\link{cond.logLik}},
  \code{\link{eff.sample.size}}, \code{\link{filter.mean}},
  \code{\link{mif2}}, \code{\link{pfilter}},
  \code{\link{pmcmc}}, \code{\link{pred.mean}}
}
\concept{particle filter methods}
