% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R
\docType{methods}
\name{Probes}
\alias{Probes}
\alias{probe}
\alias{probe,missing-method}
\alias{probe,ANY-method}
\alias{probe-pomp}
\alias{probe,pomp-method}
\alias{probe-probed_pomp}
\alias{probe,probed_pomp-method}
\title{Probes (AKA summary statistics)}
\usage{
\S4method{probe}{pomp}(object, probes, params, nsim, seed = NULL, ...,
  verbose = getOption("verbose", FALSE))

\S4method{probe}{probed_pomp}(object, probes, nsim, seed = NULL, ...,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class the extends \sQuote{pomp}}

\item{probes}{a single probe or a list of one or more probes.
A probe is simply a scalar- or vector-valued function of one argument that can be applied to the data array of a \sQuote{pomp}.
A vector-valued probe must always return a vector of the same size.
A number of useful probes are provided with the package:
see \link{Basic probes}.}

\item{params}{optional named numeric vector of model parameters.
By default, \code{params=coef(object)}.}

\item{nsim}{the number of model simulations to be computed.}

\item{seed}{optional integer;
if non-\code{NULL}, the random number generator will be initialized with this seed for simulations.
See \code{\link[=simulate-pomp]{simulate}}.}

\item{\dots}{Additional arguments.  In the case of \code{probe} and
\code{probe.match.objfun}, additional arguments are passed to
\code{\link{pomp}}, allowing one to supply new or modify existing model
characteristics or components.}

\item{verbose}{logical; print diagnostic messages?}

\item{start}{named numeric vector; the initial guess of parameters.}

\item{est}{character vector; the names of parameters to be estimated.}

\item{method}{Optimization method.  Choices refer to algorithms used in
\code{\link{optim}}, \code{\link[subplex]{subplex}}, and
\code{\link[nloptr]{nloptr}}.}

\item{fail.value}{optional numeric scalar; if non-\code{NA}, this value is
substituted for non-finite values of the objective function.  It should be
a large number (i.e., bigger than any legitimate values the objective
function is likely to take).}

\item{transform}{logical; if \code{TRUE}, optimization is performed on the
transformed scale.}
}
\value{
\code{probe} returns an object of class \sQuote{probed_pomp}, which
is derived from the \sQuote{pomp} class and contains additional information
about the \code{probe} calculation.

This information can be summarized via a call to \code{summary} and
displayed graphically via a call to \code{plot}.

\code{probe.match} returns an object of class \sQuote{probe_matched_pomp},
which is derived from class \sQuote{probed_pomp}.  It therefore contains
the information described above as well as information on the convergence
of the optimization algorithm.  This can be displayed via a call to
\code{summary}.

\code{probe.match.objfun} returns a function suitable for use as an
objective function in an \code{\link{optim}}-like optimizer.
}
\description{
Probe a partially-observed Markov process by computing summary statistics
and the synthetic likelihood.
}
\details{
\code{probe} applies one or more \dQuote{probes} to time series data and
model simulations and compares the results.  It can be used to diagnose
goodness of fit and/or as the basis for \dQuote{probe-matching}, a
generalized method-of-moments approach to parameter estimation.
\code{probe.match} calls an optimizer to adjust model parameters to do
probe-matching, i.e., to minimize the discrepancy between simulated and
actual data.  This discrepancy is measured using the \dQuote{synthetic
likelihood} as defined by Wood (2010).  \code{probe.match.objfun}
constructs an objective function for probe-matching suitable for use in
\code{optim}-like optimizers.

A call to \code{probe} results in the evaluation of the probe(s) in
\code{probes} on the data.  Additionally, \code{nsim} simulated data sets
are generated (via a call to \code{\link[=simulate-pomp]{simulate}}) and
the probe(s) are applied to each of these.  The results of the probe
computations on real and simulated data are stored in an object of class
\sQuote{probed_pomp}.

A call to \code{probe.match} results in an attempt to optimize the
agreement between model and data, as measured by the specified probes, over
the parameters named in \code{est}.  The results, including coefficients of
the fitted model and values of the probes for data and fitted-model
simulations, are stored in an object of class \sQuote{probe_matched_pomp}.

The objective function minimized by \code{probe.match}---in a form suitable
for use with \code{\link{optim}}-like optimizers---is created by a call to
\code{probe.match.objfun}.  Specifically, \code{probe.match.objfun} will
return a function that takes a single numeric-vector argument that is
assumed to cotain the parameters named in \code{est}, in that order.  This
function will return the negative synthetic log likelihood for the probes
specified.

In the case of \code{probe.match}, these are passed to the optimizer
algorithm.  These are passed via the optimizer's \code{control} list (in
the case of \code{optim}, \code{subplex}, and \code{sannbox}) or the
\code{opts} list (in the case of \code{nloptr}).
}
\section{Methods}{

\describe{
\item{plot}{ displays diagnostic plots.  }
\item{summary}{ displays summary information.  The summary includes
quantiles (fractions of simulations with probe values less than those
realized on the data) and the corresponding two-sided p-values.  In
addition, the \dQuote{synthetic likelihood} (Wood 2010) is computed, under
the assumption that the probe values are multivariate-normally distributed.
}
\item{probevals}{ extracts the realized values of the probes on the data
and on the simulations.  These are returned in a list of two elements,
\code{datvals} and \code{simvals}.  }
\item{logLik}{ returns the synthetic
likelihood for the probes.  NB: in general, this is not the same as the
likelihood.  }
\item{as.data.frame, as(object,"data.frame")}{ coercing a
\sQuote{probed_pomp} to a \sQuote{data.frame}, gives the realized values of
the probes on the data and on the simulations.  The variable \code{.id}
indicates whether the probes are from the data or simulations.  }
}
}

\references{
B. E. Kendall, C. J. Briggs, W. M. Murdoch, P. Turchin, S. P.
Ellner, E. McCauley, R. M. Nisbet, S. N. Wood Why do populations cycle? A
synthesis of statistical and mechanistic modeling approaches, Ecology,
80:1789--1805, 1999.

S. N. Wood Statistical inference for noisy nonlinear ecological dynamic
systems, Nature, 466: 1102--1104, 2010.
}
\seealso{
\link{Basic probes}, \code{\link{spect}}, and the
tutorials on the \href{https://kingaa.github.io/pomp/}{package website}.

Other summary statistics: \code{\link{Basic probes}},
  \code{\link{Probe values}}, \code{\link{abc}}
}
\author{
Daniel C. Reuman, Aaron A. King
}
\concept{summary statistics}
