% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R, R/probe_match.R
\docType{methods}
\name{Probes}
\alias{Probes}
\alias{probe}
\alias{probe,missing-method}
\alias{probe,ANY-method}
\alias{probe-pomp}
\alias{probe,pomp-method}
\alias{probe-probed_pomp}
\alias{probe,probed_pomp-method}
\alias{probe-probe_match_obfjun}
\alias{probe,probe_match_objfun-method}
\title{Probes (AKA summary statistics)}
\usage{
\S4method{probe}{pomp}(object, probes, params, nsim, seed = NULL, ...,
  verbose = getOption("verbose", FALSE))

\S4method{probe}{probed_pomp}(object, probes, nsim, seed = NULL, ...,
  verbose = getOption("verbose", FALSE))

\S4method{probe}{probe_match_objfun}(object, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class the extends \sQuote{pomp}}

\item{probes}{a single probe or a list of one or more probes.
A probe is simply a scalar- or vector-valued function of one argument that can be applied to the data array of a \sQuote{pomp}.
A vector-valued probe must always return a vector of the same size.
A number of useful probes are provided with the package:
see \link{Basic probes}.}

\item{params}{optional named numeric vector of model parameters.
By default, \code{params=coef(object)}.}

\item{nsim}{the number of model simulations to be computed.}

\item{seed}{optional integer;
if non-\code{NULL}, the random number generator will be initialized with this seed for simulations.
See \code{\link[=simulate-pomp]{simulate}}.}

\item{\dots}{Additional arguments.  In the case of \code{probe} and
\code{probe.match.objfun}, additional arguments are passed to
\code{\link{pomp}}, allowing one to supply new or modify existing model
characteristics or components.}

\item{verbose}{logical; print diagnostic messages?}
}
\value{
\code{probe} returns an object of class \sQuote{probed_pomp}, which
is derived from the \sQuote{pomp} class and contains additional information
about the \code{probe} calculation.

This information can be summarized via a call to \code{summary} and
displayed graphically via a call to \code{plot}.
}
\description{
Probe a partially-observed Markov process by computing summary statistics
and the synthetic likelihood.
}
\details{
\code{probe} applies one or more \dQuote{probes} to time series data and
model simulations and compares the results.  It can be used to diagnose
goodness of fit and/or as the basis for \dQuote{probe-matching}, a
generalized method-of-moments approach to parameter estimation.

A call to \code{probe} results in the evaluation of the probe(s) in
\code{probes} on the data.  Additionally, \code{nsim} simulated data sets
are generated (via a call to \code{\link[=simulate-pomp]{simulate}}) and
the probe(s) are applied to each of these.  The results of the probe
computations on real and simulated data are stored in an object of class
\sQuote{probed_pomp}.
}
\section{Methods}{

\describe{
\item{plot}{ displays diagnostic plots.  }
\item{summary}{ displays summary information.  The summary includes
quantiles (fractions of simulations with probe values less than those
realized on the data) and the corresponding two-sided p-values.  In
addition, the \dQuote{synthetic likelihood} (Wood 2010) is computed, under
the assumption that the probe values are multivariate-normally distributed.
}
\item{probevals}{ extracts the realized values of the probes on the data
and on the simulations.  These are returned in a list of two elements,
\code{datvals} and \code{simvals}.  }
\item{logLik}{ returns the synthetic
likelihood for the probes.  NB: in general, this is not the same as the
likelihood.  }
\item{as.data.frame, as(object,"data.frame")}{ coercing a
\sQuote{probed_pomp} to a \sQuote{data.frame}, gives the realized values of
the probes on the data and on the simulations.  The variable \code{.id}
indicates whether the probes are from the data or simulations.  }
}
}

\references{
B. E. Kendall, C. J. Briggs, W. M. Murdoch, P. Turchin, S. P.
Ellner, E. McCauley, R. M. Nisbet, S. N. Wood Why do populations cycle? A
synthesis of statistical and mechanistic modeling approaches, Ecology,
80:1789--1805, 1999.

S. N. Wood Statistical inference for noisy nonlinear ecological dynamic
systems, Nature, 466: 1102--1104, 2010.
}
\seealso{
\link{Basic probes}, \code{\link{spect}}, and the
tutorials on the \href{https://kingaa.github.io/pomp/}{package website}.

Other summary statistics: \code{\link{Basic probes}},
  \code{\link{Probe matching}}, \code{\link{Probe values}},
  \code{\link{abc}}
}
\author{
Daniel C. Reuman, Aaron A. King
}
\concept{summary statistics}
