% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe_match.R
\docType{methods}
\name{Probe matching}
\alias{Probe matching}
\alias{probe.match}
\alias{probe.match.objfun}
\alias{probe.match.objfun,missing-method}
\alias{probe.match.objfun,ANY-method}
\alias{probe.match.objfun-pomp}
\alias{probe.match.objfun,pomp-method}
\alias{probe.match.objfun-probed_pomp}
\alias{probe.match.objfun,probed_pomp-method}
\title{Probe matching}
\usage{
\S4method{probe.match.objfun}{pomp}(object, params, est, probes, nsim,
  seed = NULL, fail.value = NA, transform = FALSE, ...)

\S4method{probe.match.objfun}{probed_pomp}(object, params, est, probes,
  nsim, seed = NULL, fail.value = NA, transform = FALSE, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class the extends \sQuote{pomp}}

\item{params}{optional named numeric vector of model parameters.
By default, \code{params=coef(object)}.}

\item{est}{character vector; the names of parameters to be estimated.}

\item{probes}{a single probe or a list of one or more probes.
A probe is simply a scalar- or vector-valued function of one argument that can be applied to the data array of a \sQuote{pomp}.
A vector-valued probe must always return a vector of the same size.
A number of useful probes are provided with the package:
see \link{Basic probes}.}

\item{nsim}{the number of model simulations to be computed.}

\item{seed}{optional integer;
if non-\code{NULL}, the random number generator will be initialized with this seed for simulations.
See \code{\link[=simulate-pomp]{simulate}}.}

\item{fail.value}{optional numeric scalar; if non-\code{NA}, this value is
substituted for non-finite values of the objective function.  It should be
a large number (i.e., bigger than any legitimate values the objective
function is likely to take).}

\item{transform}{logical; if \code{TRUE}, optimization is performed on the
transformed scale.}

\item{...}{Additional arguments.  In the case of \code{probe} and
\code{probe.match.objfun}, additional arguments are passed to
\code{\link{pomp}}, allowing one to supply new or modify existing model
characteristics or components.}
}
\value{
\code{probe.match.objfun} returns an object of class \sQuote{probe_match_objfun}, which is a stateful function suitable for use as an
objective function in an \code{\link{optim}}-like optimizer.
}
\description{
Estimation of parameters by maximum synthetic likelihood
}
\details{
In probe-matching, one attempts to minimize the discrepancy between simulated and actual data, as measured by a set of summary statistics called \emph{probes}.
In \pkg{pomp}, this discrepancy is measured using the \dQuote{synthetic
likelihood} as defined by Wood (2010).

\code{probe.match.objfun} construct a stateful objective function for probe-matching suitable for use in \code{optim}-like optimizers.
Specifically, \code{probe.match.objfun} will
return a function that takes a single numeric-vector argument that is
assumed to contain the parameters named in \code{est}, in that order.
This function will return the negative synthetic log likelihood for the probes specified.
The function is stateful: each time it is called, it will remember the values of the parameters and its estimate of the synthetic likelihood.
}
\seealso{
Other summary statistics: \code{\link{Basic probes}},
  \code{\link{Probe values}}, \code{\link{Probes}},
  \code{\link{abc}}
}
\concept{summary statistics}
\keyword{optimize}
