% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstop.R
\name{pStop}
\alias{pStop}
\alias{pStop_}
\alias{pWarn}
\alias{pWarn_}
\alias{pMess}
\alias{pMess_}
\title{pStop, pWarn, pMess}
\usage{
pStop(..., who = -1L)

pStop_(...)

pWarn(..., who = -1L)

pWarn_(...)

pMess(..., who = -1L)

pMess_(...)
}
\arguments{
\item{\dots}{message}

\item{who}{integer or character.
If \code{who} is an integer, it is passed to \code{\link{sys.call}} to retrieve the name of the calling function.
One can also pass the name of the calling function in \code{who}.
In either case, the name of the calling function is included in the message.}
}
\description{
Custom error, warning, and message functions.
}
\keyword{internal}
