% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{melt}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reshape2}{\code{\link[reshape2]{melt}}}
}}

