% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{systematic_resample}
\title{Resample}
\usage{
systematic_resample(weights, Np = length(weights))
}
\arguments{
\item{weights}{numeric; vector of weights.}

\item{np}{integer scalar; number of samples to draw.}
}
\value{
A vector of integers containing the indices of the resample.
}
\description{
Systematic resampling.
}
\concept{sampling}
\keyword{internal}
