% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\docType{data}
\name{ricker}
\alias{ricker}
\title{Ricker model with Poisson observations.}
\description{
\code{ricker} is a \sQuote{pomp} object encoding a stochastic Ricker model
with Poisson measurement error.
}
\details{
The state process is \eqn{N_{t+1} = r N_{t} \exp(-c N_{t}+e_{t})}{N[t+1] =
r N[t] exp(-c N[t]+e[t])}, where the \eqn{e_t}{e[t]} are i.i.d. normal
random deviates with zero mean and variance \eqn{\sigma^2}{sigma^2}.  The
observed variables \eqn{y_t}{y[t]} are distributed as
\eqn{\mathrm{Poisson}(\phi N_t)}{Poisson(phi N[t])}.
}
\examples{

pompExample(ricker)
plot(ricker)
coef(ricker)

}
\seealso{
Other pomp examples: \code{\link{blowflies}},
  \code{\link{dacca}}, \code{\link{gompertz}},
  \code{\link{measles}}, \code{\link{ou2}},
  \code{\link{pompExample}}, \code{\link{rw2}},
  \code{\link{sir_models}}
}
\concept{pomp examples}
\keyword{datasets}
\keyword{models}
