% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinit_spec.R
\name{rinit_spec}
\alias{rinit_spec}
\title{The initial-state distribution}
\description{
Specification of rinit
}
\details{
To fully specify the unobserved Markov state process, one must give its distribution at the zero-time (\code{t0}).

By default, \code{pomp} assumes that this initial distribution is concentrated on a single point.
In particular, any parameters in \code{params}, the names of which end in \dQuote{\code{.0}}, are assumed to be initial values of states.
When the state process is initialized, these are simply copied over as initial conditions.
The names of the resulting state variables are obtained by dropping the \dQuote{\code{.0}} suffix.

One can override this default behavior by furnishing a value for the \code{rinit} argument of most \pkg{pomp} methods.
As usual, this can be provided either as a C snippet or as an \R function.
In the former case, bear in mind that:
\enumerate{
  \item The goal of a this snippet is the construction of a state vector, i.e., the setting of the dynamical states at time \eqn{t_0}{\code{t0}}.
  \item In addition to the parameters and covariates (if any), the variable \code{t}, containing the zero-time, will be defined in the context in which the snippet is executed.
  \item \strong{NB:} The \code{statenames} argument plays a particularly important role when the rinit is specified using a C snippet.
   In particular, every state variable must be named in \code{statenames}.
   \strong{Failure to follow this rule will result in undefined behavior.}
 }

If an \R function is to be used, pass \preformatted{
   rinit = f}
to \code{pomp}, where \code{f} is a function with prototype \preformatted{
   f(params, t0, \dots)}
When \code{f} is called,
\itemize{
  \item \code{params} will be a named numeric vector of parameters.
  \item \code{t0} will be the time at which initial conditions are desired.
}
As usual, \code{f} may take additional arguments, provided these are passed along with it in the call to \code{pomp}.
\code{f} must return a named numeric vector of initial states.
It is of course important that the names of the states match the expectations of the other basic components.

Note that the state-process \code{rinit} can be either deterministic (the default) or stochastic.
In the latter case, it samples from the distribution of the state process at the zero-time, \code{t0}.
}
\seealso{
Other information on model implementation: \code{\link{Csnippet}},
  \code{\link{accumulators}},
  \code{\link{covariate_table}},
  \code{\link{dmeasure_spec}}, \code{\link{dprocess_spec}},
  \code{\link{parameter_trans}}, \code{\link{prior_spec}},
  \code{\link{rmeasure_spec}}, \code{\link{rprocess_spec}},
  \code{\link{skeleton_spec}}, \code{\link{userdata}}
}
\concept{information on model implementation}
