% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\docType{methods}
\name{rmeasure}
\alias{rmeasure}
\alias{rmeasure,pomp-method}
\alias{rmeasure-pomp}
\title{rmeasure}
\usage{
\S4method{rmeasure}{pomp}(object, x, times, params, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}.}

\item{x}{a rank-3 array containing states of the unobserved process.
The dimensions of \code{x} are \code{nvars} x \code{nrep} x \code{ntimes},
where \code{nvars} is the number of state variables,
\code{nrep} is the number of replicates,
and \code{ntimes} is the length of \code{times}.}

\item{times}{a numeric vector (length \code{ntimes}) containing times.
These must be in non-decreasing order.}

\item{params}{a \code{npar} x \code{nrep} matrix of parameters.
Each column is an independent parameter set and is paired with the corresponding column of \code{x} or \code{xstart}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{rmeasure} returns a rank-3 array of dimensions \code{nobs} x \code{nrep} x \code{ntimes}, where \code{nobs} is the number of observed variables.
}
\description{
Sample from the measurement model distribution, given values of the latent states and the parameters.
}
\seealso{
Other pomp workhorses: \code{\link{dmeasure}},
  \code{\link{dprior}}, \code{\link{dprocess}},
  \code{\link{partrans}}, \code{\link{rinit}},
  \code{\link{rprior}}, \code{\link{rprocess}},
  \code{\link{skeleton}}, \code{\link{workhorses}}
}
\concept{pomp workhorses}
\keyword{internal}
