% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\docType{methods}
\name{rprior}
\alias{rprior}
\alias{rprior,pomp-method}
\alias{rprior-pomp}
\alias{rprior,ANY-method}
\alias{rprior,missing-method}
\title{rprior}
\usage{
\S4method{rprior}{pomp}(object, params, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}.
This will typically be the output of \code{pomp}, \code{simulate}, or one of the \pkg{pomp} inference algorithms.}

\item{params}{a \code{npar} x \code{nrep} matrix of parameters.
Each column is treated as an independent parameter set, in correspondence with the corresponding column of \code{x}.}

\item{...}{additional arguments are ignored.}
}
\value{
A numeric matrix containing the required samples.
}
\description{
Sample from the prior probability distribution.
}
\seealso{
Specification of the prior distribution simulator: \code{\link{prior_spec}}

Other pomp workhorses: 
\code{\link{dmeasure}()},
\code{\link{dprior}()},
\code{\link{dprocess}()},
\code{\link{flow}()},
\code{\link{partrans}()},
\code{\link{rinit}()},
\code{\link{rmeasure}()},
\code{\link{rprocess}()},
\code{\link{skeleton}()},
\code{\link{workhorses}}
}
\concept{pomp workhorses}
