% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw2.R
\docType{data}
\name{rw2}
\alias{rw2}
\title{Two-dimensional random-walk process}
\description{
\code{rw2} is a \sQuote{pomp} object encoding a 2-D normal random walk.
}
\details{
The random-walk process is fully but noisily observed.
}
\examples{

pompExample(rw2)
plot(rw2)
x <- simulate(rw2,nsim=10,seed=20348585L,params=c(x1.0=0,x2.0=0,s1=1,s2=3,tau=1))
plot(x[[1]])

}
\seealso{
Other pomp examples: \code{\link{blowflies}},
  \code{\link{dacca}}, \code{\link{gompertz}},
  \code{\link{measles}}, \code{\link{ou2}},
  \code{\link{pompExample}}, \code{\link{ricker}},
  \code{\link{sir_models}}
}
\concept{pomp examples}
\keyword{datasets}
\keyword{models}
