% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw2.R
\docType{data}
\name{rw2}
\alias{rw2}
\title{Two-dimensional random-walk process}
\usage{
rw2(
  x1_0 = 0,
  x2_0 = 0,
  s1 = 1,
  s2 = 3,
  tau = 1,
  times = 1:100,
  t0 = 0,
  seed = 1376784970L
)
}
\arguments{
\item{x1_0, x2_0}{initial conditions (i.e., latent state variable values at the zero time \code{t0})}

\item{s1, s2}{random walk intensities}

\item{tau}{observation error s.d.}

\item{times}{observation times}

\item{t0}{zero time}

\item{seed}{seed of the random number generator}
}
\value{
A \sQuote{pomp} object containing simulated data.
}
\description{
\code{rw2} constructs a \sQuote{pomp} object encoding a 2-D Gaussian random walk.
}
\details{
The random-walk process is fully but noisily observed.
}
\examples{
\donttest{

  if (require(ggplot2)) {

    rw2() |> plot()

    rw2(s1=1,s2=1,tau=0.1) |>
      simulate(nsim=10,format="d") |>
      ggplot(aes(x=y1,y=y2,group=.id,color=.id))+
      geom_path()+
      guides(color="none")+
      theme_bw()

  }
}
}
\seealso{
More examples provided with \pkg{pomp}: 
\code{\link{blowflies}},
\code{\link{childhood_disease_data}},
\code{\link{compartmental_models}},
\code{\link{dacca}()},
\code{\link{ebola}},
\code{\link{gompertz}()},
\code{\link{ou2}()},
\code{\link{pomp_examples}},
\code{\link{ricker}()},
\code{\link{verhulst}()}
}
\concept{pomp examples}
\keyword{models}
