% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\docType{methods}
\name{simulate}
\alias{simulate}
\alias{simulate-missing}
\alias{simulate,missing-method}
\alias{simulate-data.frame}
\alias{simulate,data.frame-method}
\alias{simulate-pomp}
\alias{simulate,pomp-method}
\title{Simulations of a partially-observed Markov process}
\usage{
\S4method{simulate}{missing}(nsim = 1, seed = NULL, rinit, rprocess,
  rmeasure, params, times, t0, format = c("pomps", "arrays",
  "data.frame"), include.data = FALSE, ...,
  verbose = getOption("verbose", FALSE))

\S4method{simulate}{data.frame}(object, nsim = 1, seed = NULL, rinit,
  rprocess, rmeasure, params, times, t0, format = c("pomps", "arrays",
  "data.frame"), include.data = FALSE, ...,
  verbose = getOption("verbose", FALSE))

\S4method{simulate}{pomp}(object, nsim = 1, seed = NULL, rinit,
  rprocess, rmeasure, params, times, t0, format = c("pomps", "arrays",
  "data.frame"), include.data = FALSE, ...,
  verbose = getOption("verbose", FALSE))
}
\arguments{
\item{nsim}{The number of simulations to perform.
Note that the number of replicates will be \code{nsim} times \code{ncol(params)}.}

\item{seed}{optional;
if set, the pseudorandom number generator (RNG) will be initialized with \code{seed}.  the random seed to use.
The RNG will be restored to its original state afterward.}

\item{rinit}{simulator of the initial-state distribution.
This can be furnished either as a C snippet, an \R function, or the name of a pre-compiled native routine available in a dynamically loaded library.
Setting \code{rinit=NULL} sets the initial-state simulator to its default.
For more information, see \link[=rinit_spec]{here}.}

\item{rprocess}{simulator of the latent state process, specified using one of the \link[=rprocess_spec]{rprocess plugins}.
Setting \code{rprocess=NULL} removes the latent-state simulator.
For more information, \link[=rprocess_spec]{see the documentation on these plugins}.}

\item{rmeasure}{simulator of the measurement model, specified either as a C snippet, an \R function, or the name of a pre-compiled native routine available in a dynamically loaded library.
Setting \code{rmeasure=NULL} removes the measurement model simulator.
For more information, see \link[=rmeasure_spec]{here}.}

\item{params}{a named numeric vector or a matrix with rownames
containing the parameters at which the simulations are to be performed.}

\item{times}{the times at which observations are made.
\code{times} must indicate the column of observation times by name or index.
The time vector must be numeric and strictly increasing.
Internally, \code{data} will be internally coerced to an array with storage-mode \code{double}.}

\item{t0}{The zero-time, i.e., the time of the initial state.
This must be no later than the time of the first observation, i.e., \code{t0 <= times[1]}.}

\item{format}{the format in which to return the results.

\code{format = "pomps"} causes the results to be returned as a single \dQuote{pomp} object,
identical to \code{object} except for the latent states and observations,
which have been replaced by the simulated values.

\code{format = "arrays"} causes the results to be returned as a list of two arrays.
The \dQuote{states} element will contain the simulate state trajectories in a rank-3 array with dimensions
\code{nvar} x \code{(ncol(params)*nsim)} x \code{ntimes}.
Here, \code{nvar} is the number of state variables and \code{ntimes} the length of the argument \code{times}.
The \dQuote{obs} element will contain the simulated data, returned as a rank-3 array with dimensions
\code{nobs} x \code{(ncol(params)*nsim)} x \code{ntimes}.
Here, \code{nobs} is the number of observables.

\code{format = "data.frame"} causes the results to be returned as a single data frame containing
the time, states, and observations.
An ordered factor variable, \sQuote{.id}, distinguishes one simulation from another.}

\item{include.data}{if \code{TRUE}, the original data are included (with \code{.id = "rep"}).
This option is ignored unless \code{format = "data.frame"}.}

\item{...}{additional arguments supply new or modify existing model characteristics or components.
See \code{\link{pomp}} for a full list of recognized arguments.

When named arguments not recognized by \code{\link{pomp}} are provided, these are made available to all basic components via the so-called \dfn{userdata} facility.
This allows the user to pass information to the basic components outside of the usual routes of covariates (\code{covar}) and model parameters (\code{params}).
See the \link[=userdata]{userdata documentation here} for information on how to use this facility.}

\item{verbose}{logical; if \code{TRUE}, diagnostic messages will be printed to the console.}

\item{object}{optional; if present, it should be the output of one of \pkg{pomp}'s methods}
}
\value{
A single \dQuote{pomp} object,
a \dQuote{pompList} object,
a named list of two arrays,
or a data frame, according to the \code{format} option.

If \code{params} is a matrix, each column is treated as a distinct parameter set.
In this case, if \code{nsim=1},
then \code{simulate} will return one simulation for each parameter set.
If \code{nsim>1},
then \code{simulate} will yield \code{nsim} simulations for each parameter set.
These will be ordered such that
the first \code{ncol(params)} simulations represent one simulation
from each of the distinct parameter sets,
the second \code{ncol(params)} simulations represent a second simulation from each,
and so on.

Adding column names to \code{params} can be helpful.
}
\description{
\code{simulate} generates simulations of the state and measurement
processes.
}
\seealso{
Other elementary POMP methods: \code{\link{pfilter}},
  \code{\link{pomp-package}}, \code{\link{probe}}
}
\author{
Aaron A. King
}
\concept{elementary POMP methods}
