% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\docType{data}
\name{sir_models}
\alias{sir_models}
\alias{sir}
\alias{sir2}
\title{Compartmental epidemiological models}
\description{
\code{sir} is a \sQuote{pomp} object encoding a simple seasonal SIR model.
Simulation is performed using an Euler multinomial approximation.

\code{sir2} has the same model implemented using Gillespie's algorithm.

This and similar examples are discussed and constructed in tutorials
available on the \href{https://kingaa.github.io/pomp/}{package website}.
}
\details{
Do \preformatted{
    pompExample(sir,show=TRUE)
    pompExample(sir2,show=TRUE)
}
to see the codes that generate these examples.
}
\examples{

pompExample(sir)
plot(sir)
plot(simulate(sir))
coef(sir)

pompExample(sir2)
plot(sir2)
plot(simulate(window(sir2,end=3)))
coef(sir2)

}
\seealso{
Other pomp examples: \code{\link{blowflies}},
  \code{\link{dacca}}, \code{\link{gompertz}},
  \code{\link{measles}}, \code{\link{ou2}},
  \code{\link{pompExample}}, \code{\link{ricker}},
  \code{\link{rw2}}
}
\concept{pomp examples}
\keyword{datasets}
\keyword{models}
