% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R, R/spect_match.R
\docType{methods}
\name{spect}
\alias{spect}
\alias{spect,missing-method}
\alias{spect,ANY-method}
\alias{spect-pomp}
\alias{spect,pomp-method}
\alias{spect-spectd_pomp}
\alias{spect,spectd_pomp-method}
\alias{spect-spect_match_objfun}
\alias{spect,spect_match_objfun-method}
\title{Power spectrum}
\usage{
\S4method{spect}{pomp}(object, params, vars, kernel.width, nsim,
  seed = NULL, transform.data = identity, detrend = c("none", "mean",
  "linear", "quadratic"), ...)

\S4method{spect}{spectd_pomp}(object, params, vars, kernel.width, nsim,
  seed = NULL, transform.data, detrend, ...)

\S4method{spect}{spect_match_objfun}(object, ...)
}
\arguments{
\item{object}{An object of class \sQuote{pomp}.}

\item{params}{optional named numeric vector of model parameters.  By
default, \code{params=coef(object)}.}

\item{vars}{optional; names of observed variables for which the power spectrum will be computed.
By default, the spectrum will be computed for all observables.}

\item{kernel.width}{width parameter for the smoothing kernel used for
calculating the estimate of the spectrum.}

\item{nsim}{number of model simulations to be computed.}

\item{seed}{optional; if non-\code{NULL}, the random number generator will
be initialized with this seed for simulations.
See \code{\link[=simulate-pomp]{simulate}}.}

\item{transform.data}{function; this transformation will be applied to the
observables prior to estimation of the spectrum, and prior to any
detrending.}

\item{detrend}{de-trending operation to perform.  Options include no
detrending, and subtraction of constant, linear, and quadratic trends from
the data.  Detrending is applied to each data series and to each model
simulation independently.}

\item{\dots}{ignored}
}
\value{
An object of class \sQuote{spectd_pomp}.
}
\description{
Power spectrum computation and spectrum-matching for partially-observed
Markov processes.
}
\details{
\code{spect} estimates the power spectrum of time series data and model
simulations and compares the results.  It can be used to diagnose goodness
of fit and/or as the basis for frequency-domain parameter estimation
(\code{spect.match}).

A call to \code{spect} results in the estimation of the power spectrum for
the (transformed, detrended) data and \code{nsim} model simulations.  The
results of these computations are stored in an object of class
\sQuote{spectd_pomp}.
}
\references{
D.C. Reuman, R.A. Desharnais, R.F. Costantino, O. Ahmad, J.E.
Cohen (2006) Power spectra reveal the influence of stochasticity on
nonlinear population dynamics.  \emph{Proceedings of the National Academy
of Sciences} \bold{103}, 18860-18865.

D.C. Reuman, R.F. Costantino, R.A. Desharnais, J.E. Cohen (2008) Color of
environmental noise affects the nonlinear dynamics of cycling,
stage-structured populations.  \emph{Ecology Letters}, \bold{11}, 820-830.
}
\seealso{
\code{\link{simulate}}, \code{\link{probe}}

Other summary statistics methods: \code{\link{abc}},
  \code{\link{basic_probes}}, \code{\link{probe.match}},
  \code{\link{probevals}}, \code{\link{probe}}
}
\author{
Daniel C. Reuman, Cai GoGwilt, Aaron A. King
}
\concept{summary statistics methods}
