% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy.R
\name{spy}
\alias{spy}
\alias{spy,missing-method}
\alias{spy,ANY-method}
\alias{spy,pomp-method}
\title{Spy}
\usage{
\S4method{spy}{pomp}(object)
}
\arguments{
\item{object}{the object whose structure we wish to examine}
}
\description{
Peek into the inside of one of \pkg{pomp}'s objects.
}
