% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy.R
\name{spy}
\alias{spy}
\alias{spy,missing-method}
\alias{spy,ANY-method}
\alias{spy,pomp-method}
\title{Spy}
\usage{
\S4method{spy}{pomp}(object)
}
\arguments{
\item{object}{the object whose structure we wish to examine}
}
\description{
Peek into the inside of one of \pkg{pomp}'s objects.
}
\examples{
\donttest{

  ricker() |> spy()

  sir() |> spy()

  sir2() |> spy()

}
}
\seealso{
Csnippet

Other extraction methods: 
\code{\link{coef}()},
\code{\link{cond_logLik}()},
\code{\link{covmat}()},
\code{\link{eff_sample_size}()},
\code{\link{filter_mean}()},
\code{\link{filter_traj}()},
\code{\link{forecast}()},
\code{\link{logLik}},
\code{\link{obs}()},
\code{\link{pred_mean}()},
\code{\link{pred_var}()},
\code{\link{saved_states}()},
\code{\link{states}()},
\code{\link{summary}()},
\code{\link{time}()},
\code{\link{timezero}()},
\code{\link{traces}()}
}
\concept{extraction methods}
