% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{methods}
\name{states}
\alias{states}
\alias{states,pomp-method}
\title{Latent states}
\usage{
\S4method{states}{pomp}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{names of variables to retrieve}

\item{...}{ignored}
}
\description{
Extract the latent states from a \sQuote{pomp} object.
}
