% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{methods}
\name{states}
\alias{states}
\alias{states,pomp-method}
\alias{states,listie-method}
\title{Latent states}
\usage{
\S4method{states}{pomp}(object, vars, ...)

\S4method{states}{listie}(object, vars, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class extending \sQuote{pomp}}

\item{vars}{names of variables to retrieve}

\item{...}{ignored}
}
\description{
Extract the latent states from a \sQuote{pomp} object.
}
\seealso{
Other extraction methods: 
\code{\link{coef}()},
\code{\link{cond.logLik}()},
\code{\link{covmat}()},
\code{\link{eff.sample.size}()},
\code{\link{filter.mean}()},
\code{\link{filter.traj}()},
\code{\link{forecast}()},
\code{\link{logLik}},
\code{\link{obs}()},
\code{\link{pred.mean}()},
\code{\link{pred.var}()},
\code{\link{saved.states}()},
\code{\link{spy}()},
\code{\link{summary}()},
\code{\link{timezero}()},
\code{\link{time}()},
\code{\link{traces}()}
}
\concept{extraction methods}
