% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R, R/probe.R, R/spect.R, R/objfun.R
\name{summary}
\alias{summary}
\alias{summary,probed_pomp-method}
\alias{summary,spectd_pomp-method}
\alias{summary,objfun-method}
\title{Summary methods}
\usage{
\S4method{summary}{probed_pomp}(object, ...)

\S4method{summary}{spectd_pomp}(object, ...)

\S4method{summary}{objfun}(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{\dots}{ignored or passed to the more primitive function}
}
\description{
Display a summary of a fitted model object.
}
\seealso{
Other extraction methods: 
\code{\link{coef}()},
\code{\link{cond.logLik}()},
\code{\link{covmat}()},
\code{\link{eff.sample.size}()},
\code{\link{filter.mean}()},
\code{\link{filter.traj}()},
\code{\link{forecast}()},
\code{\link{logLik}},
\code{\link{obs}()},
\code{\link{pred.mean}()},
\code{\link{pred.var}()},
\code{\link{saved.states}()},
\code{\link{spy}()},
\code{\link{states}()},
\code{\link{timezero}()},
\code{\link{time}()},
\code{\link{traces}()}
}
\concept{extraction methods}
