% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,probed_pomp-method}
\alias{summary,spectd_pomp-method}
\alias{summary,probe_matched_pomp-method}
\alias{summary,spect_matched_pomp-method}
\alias{summary,traj_matched_pomp-method}
\alias{summary-probed_pomp}
\alias{summary-spectd_pomp}
\alias{summary-traj_matched_pomp}
\alias{summary-spect_matched_pomp}
\alias{summary-probe_matched_pomp}
\title{Summary methods}
\usage{
\S4method{summary}{probed_pomp}(object, ...)

\S4method{summary}{spectd_pomp}(object, ...)

\S4method{summary}{traj_matched_pomp}(object, ...)

\S4method{summary}{spect_matched_pomp}(object, ...)

\S4method{summary}{probe_matched_pomp}(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{\dots}{ignored}
}
\description{
Display a summary of a fitted model object.
}
