% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R, R/spect.R, R/summary.R,
%   R/probe_match.R, R/spect_match.R
\docType{methods}
\name{summary-probed_pomp}
\alias{summary-probed_pomp}
\alias{summary,probed_pomp-method}
\alias{summary-spectd_pomp}
\alias{summary,spectd_pomp-method}
\alias{summary}
\alias{summary-probe_match_obfjun}
\alias{summary,probe_match_objfun-method}
\alias{summary-spect_match_objfun}
\alias{summary,spect_match_objfun-method}
\title{Summary methods}
\usage{
\S4method{summary}{probed_pomp}(object)

\S4method{summary}{spectd_pomp}(object)

\S4method{summary}{probe_match_objfun}(object)

\S4method{summary}{spect_match_objfun}(object)
}
\arguments{
\item{object}{a fitted model object}

\item{\dots}{ignored}
}
\description{
Display a summary of a fitted model object.
}
