% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R, R/spect.R, R/summary.R, R/objfun.R
\docType{methods}
\name{summary-probed_pomp}
\alias{summary-probed_pomp}
\alias{summary,probed_pomp-method}
\alias{summary-spectd_pomp}
\alias{summary,spectd_pomp-method}
\alias{summary}
\alias{summary-objfun}
\alias{summary,objfun-method}
\title{Summary methods}
\usage{
\S4method{summary}{probed_pomp}(object)

\S4method{summary}{spectd_pomp}(object)

\S4method{summary}{objfun}(object)
}
\arguments{
\item{object}{a fitted model object}

\item{\dots}{ignored}
}
\description{
Display a summary of a fitted model object.
}
