% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R, R/probe.R, R/spect.R, R/objfun.R
\name{summary}
\alias{summary}
\alias{summary,probed_pomp-method}
\alias{summary,spectd_pomp-method}
\alias{summary,objfun-method}
\title{Summary methods}
\usage{
\S4method{summary}{probed_pomp}(object, ...)

\S4method{summary}{spectd_pomp}(object, ...)

\S4method{summary}{objfun}(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{\dots}{ignored}
}
\description{
Display a summary of a fitted model object.
}
