% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezero.R
\docType{methods}
\name{timezero}
\alias{timezero}
\alias{timezero<-}
\alias{timezero,missing-method}
\alias{timezero,ANY-method}
\alias{timezero<-,missing-method}
\alias{timezero<-,ANY-method}
\alias{timezero,pomp-method}
\alias{timezero<-,pomp-method}
\title{The zero time}
\usage{
\S4method{timezero}{pomp}(object, ...)

\S4method{timezero}{pomp}(object, ...) <- value
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}}

\item{\dots}{ignored or passed to the more primitive function}

\item{value}{numeric; the new zero-time value}
}
\value{
the value of the zero time
}
\description{
Get and set the zero-time.
}
\seealso{
Other extraction methods: 
\code{\link{coef}()},
\code{\link{cond.logLik}()},
\code{\link{covmat}()},
\code{\link{eff.sample.size}()},
\code{\link{filter.mean}()},
\code{\link{filter.traj}()},
\code{\link{forecast}()},
\code{\link{logLik}},
\code{\link{obs}()},
\code{\link{pred.mean}()},
\code{\link{pred.var}()},
\code{\link{saved.states}()},
\code{\link{spy}()},
\code{\link{states}()},
\code{\link{summary}()},
\code{\link{time}()},
\code{\link{traces}()}
}
\concept{extraction methods}
