% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezero.R
\docType{methods}
\name{timezero}
\alias{timezero}
\alias{timezero<-}
\alias{timezero,missing-method}
\alias{timezero,ANY-method}
\alias{timezero<-,missing-method}
\alias{timezero<-,ANY-method}
\alias{timezero,pomp-method}
\alias{timezero<-,pomp-method}
\title{The zero time}
\usage{
\S4method{timezero}{pomp}(object, ...)

\S4method{timezero}{pomp}(object, ...) <- value
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class that extends \sQuote{pomp}}

\item{\dots}{ignored}

\item{value}{numeric; the new zero-time value}
}
\value{
the value of the zero time
}
\description{
Get and set the zero-time.
}
