% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj_match.R
\docType{methods}
\name{traj.match}
\alias{traj.match}
\alias{traj.match.objfun,missing-method}
\alias{traj.match.objfun,ANY-method}
\alias{traj.match.objfun-pomp}
\alias{traj.match.objfun}
\alias{traj.match.objfun,pomp-method}
\title{Trajectory matching}
\usage{
\S4method{traj.match.objfun}{pomp}(object, params, est,
  transform = FALSE, ...)
}
\arguments{
\item{object}{an object of class \sQuote{pomp}, or of a class the extends \sQuote{pomp}}

\item{params}{optional named numeric vector of model parameters.
By default, \code{params=coef(object)}.}

\item{est}{character vector; the names of parameters to be estimated.}

\item{transform}{logical;
if \code{TRUE}, optimization is to be performed on the transformed scale.}

\item{...}{additional arguments which can be used to change the defaults or modify
model characteristics or components.}
}
\value{
\code{traj.match.objfun} returns a function suitable for use as an
objective function in an \code{\link{optim}}-like optimizer.
}
\description{
Fitting a deterministic model to data.
}
\details{
In trajectory matching, one attempts to minimize the discrepancy between a \acronym{POMP} model's predictions and data under the assumption that the process model is deterministic and all discrepancies between model and data are due to measurement error.
The measurement model likelihood (evaluated by \code{dmeasure}), or rather its negative, is a natural measure of the discrepancy.

Thus trajectory matching is a generalization of the traditional nonlinear least squares approach.
In particular, if, on some scale, the measurement errors are normal with constant variance, then trajectory matching is equivalent to least squares (on that particular scale).

The objective function itself --- in a form suitable for use with \code{\link{optim}}-like optimizers --- is created by a call to \code{traj.match.objfun}.
Specifically, \code{traj.match.objfun} will return a function that takes a single numeric-vector argument that is assumed to cotain the parameters named in \code{est}, in that order.
}
\seealso{
\code{\link{trajectory}}, \code{\link{optim}},
\code{\link[subplex]{subplex}}, \code{\link[nloptr]{nloptr}}
}
