% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj_match.R
\docType{methods}
\name{Trajectory matching}
\alias{Trajectory matching}
\alias{traj.match.objfun,pomp-method}
\alias{traj.match,pomp-method}
\alias{traj.match,traj_matched_pomp-method}
\title{Trajectory matching}
\usage{
\S4method{traj.match.objfun}{pomp}(object, params, est,
  transform = FALSE, ...)

\S4method{traj.match}{pomp}(object, start, est = character(0),
  method = c("Nelder-Mead", "subplex", "SANN", "BFGS", "sannbox",
  "nloptr"), transform = FALSE, ...)

\S4method{traj.match}{traj_matched_pomp}(object, est, transform, ...)
}
\arguments{
\item{object}{A \sQuote{\link{pomp}} object.  If \code{object} has no
\code{skeleton} slot, an error will be generated.}

\item{params}{optional named numeric vector of parameters.  This should
contain all parameters needed by the \code{skeleton} and \code{dmeasure}
slots of \code{object}.  In particular, any parameters that are to be
treated as fixed should be present here.  Parameter values given in
\code{params} for parameters named in \code{est} will be ignored.  By
default, \code{params=coef(object)} if the latter exists.}

\item{est}{character vector containing the names of parameters to be
estimated.  In the case of \code{traj.match.objfun}, the objective function
that is constructed will assume that its argument contains the parameters
in this order.}

\item{transform}{logical; if \code{TRUE}, optimization is performed on the
transformed scale.}

\item{\dots}{Extra arguments that will be passed either to the optimizer
(\code{\link{optim}}, \code{\link[subplex]{subplex}},
\code{\link[nloptr]{nloptr}}, or \code{\link{sannbox}}, via their
\code{control} (\code{optim}, \code{subplex}, \code{sannbox}) or
\code{opts} (\code{nloptr}) lists) or to the ODE integrator.  In
\code{traj.match}, extra arguments will be passed to the optimizer.  In
\code{traj.match.objfun}, extra arguments are passed to
\code{\link{trajectory}}.  If extra arguments are needed by both optimizer
and \code{\link{trajectory}}, construct an objective function first using
\code{traj.match.objfun}, then give this objective function to the
optimizer.}

\item{start}{named numeric vector containing an initial guess for
parameters.  By default \code{start=coef(object)} if the latter exists.}

\item{method}{Optimization method.  Choices are
\code{\link[subplex]{subplex}}, \dQuote{sannbox}, any of the methods used
by \code{\link{optim}}, and \code{nloptr}.  The latter makes available all
the optimization algorithms of the \pkg{nloptr} package
(\url{https://cran.r-project.org/package=nloptr}).}
}
\value{
\code{traj.match} returns an object of class
\sQuote{traj_matched_pomp}.  Available methods for objects of this class
include \code{summary} and \code{logLik}.

\code{traj.match.objfun} returns a function suitable for use as an
objective function in an \code{\link{optim}}-like optimizer.
}
\description{
Fitting a deterministic model to data.
}
\seealso{
\code{\link{trajectory}}, \code{\link{optim}},
\code{\link[subplex]{subplex}}
}
