% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\docType{methods}
\name{workhorses}
\alias{workhorses}
\title{Workhorse functions for the \pkg{pomp} algorithms.}
\description{
These functions mediate the interface between the user's model and the package algorithms.
They are low-level functions that do the work needed by the package's inference methods.
}
\details{
They include \describe{
\item{\code{\link{dmeasure}}}{which evaluates the measurement model density,}
\item{\code{\link{rmeasure}}}{which samples from the measurement model distribution,}
\item{\code{\link{dprocess}}}{which evaluates the process model density,}
\item{\code{\link{rprocess}}}{which samples from the process model distribution,}
\item{\code{\link{dprior}}}{which evaluates the prior probability density,}
\item{\code{\link{rprior}}}{which samples from the prior distribution,}
\item{\code{\link{skeleton}}}{which evaluates the model's deterministic skeleton,}
\item{\code{\link{flow}}}{which iterates or integrates the deterministic skeleton to yield trajectories,}
\item{\code{\link{partrans}}}{which performs parameter transformations associated with the model.}
}
}
\seealso{
\code{\link[=simulate-pomp]{simulate}}, \code{\link{trajectory}},
\code{\link{pfilter}}, \code{\link{probe}}

Other pomp workhorses: \code{\link{dmeasure}},
  \code{\link{dprior}}, \code{\link{dprocess}},
  \code{\link{flow}}, \code{\link{partrans}},
  \code{\link{rinit}}, \code{\link{rmeasure}},
  \code{\link{rprior}}, \code{\link{rprocess}},
  \code{\link{skeleton}}
}
\author{
Aaron A. King
}
\concept{extending the pomp package}
\concept{low-level interface}
\concept{pomp workhorses}
