% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workhorses.R
\docType{methods}
\name{workhorses}
\alias{workhorses}
\title{Workhorse functions for the \pkg{pomp} algorithms.}
\description{
These functions mediate the interface between the user's model and the package algorithms.
They are low-level functions that do the work needed by the package's inference methods.
}
\details{
They include \describe{
\item{dmeasure}{which evaluates the measurement model density,}
\item{rmeasure}{which samples from the measurement model distribution,}
\item{dprocess}{which evaluates the process model density,}
\item{rprocess}{which samples from the process model distribution,}
\item{dprior}{which evaluates the prior probability density,}
\item{rprior}{which samples from the prior distribution,}
\item{skeleton}{which evaluates the model's deterministic skeleton,}
\item{partrans}{which performs parameter transformations associated with the model.}
}
}
\seealso{
\code{\link[=simulate-pomp]{simulate}} \code{\link{trajectory}}

Other pomp workhorses: \code{\link{dmeasure}},
  \code{\link{dprior}}, \code{\link{dprocess}},
  \code{\link{partrans}}, \code{\link{rinit}},
  \code{\link{rmeasure}}, \code{\link{rprior}},
  \code{\link{rprocess}}, \code{\link{skeleton}}
}
\author{
Aaron A. King
}
\concept{pomp workhorses}
\keyword{internal}
\keyword{programming}
